### 2019-10-24 #347 @MohammedFCIS
submission: dbparser
Submitting Author: Mohammed Ali(@submission:)
Repository: https://github.com/Dainanahan/dbparser
Version submitted:   v1.0.4
Version accepted: TBD

---

```
Package: dbparser
Title: 'DrugBank' Database XML Parser
Version: 1.0.4
    c(
    person("Mohammed", "Ali", email = "moh_fcis@yahoo.com", role = c("aut", "cre")),
    person("Ali", "Ezzat", email = "ali_ezzat85@yahoo.com", role = "aut"))
Description: This tool is for parsing the 'DrugBank' XML database <http://drugbank.ca/>. The parsed
    data are then returned in a proper 'R' dataframe with the ability to save
    them in a given database.
    DBI,
    odbc,
    RMariaDB,
    tools,
    XML

```

	- [ ] workflow automataion

The package accesses DrugBank DB, extract and transform data into R tibbles, and provides a set of functions to save them in a given SQL database.

Pharmaceutical researches who are interested to use data science and machine learning tools in their Drugs Development research

-   Are there other R packages that accomplish the same thing? If so, how does yours differ or meet
There is **drugbankR** that might do a similar task by loading the whole DrugBank database into a single R dataframe and provide some queries to for specific data.
On the other hand, **dbparser** extract drugbank database in to different 75 scientifically detailed dataframe with clear attributes. These dataframes need no extra functionalities to query them except regular base R functionalities. Also these dataframe can be saved in to a given SQL database.
Also, **dbparser** comes to greater coverage in terms of functionality, documentation, and robust test coverage and error handling, In comparison to **drugbankR**.

- [x] Do you intend for this package to go on CRAN?  (it is already on CRAN)
