### 2019-11-02 31f04f0
#70, add code of conduct to read me
