+++ README.Rmd
@@ -158,3 +158,8 @@ p <-
 ## display plot
 p
 ```
+
+## Code of Conduct
+Please note that the 'dbparser' project is released with a
+  [Contributor Code of Conduct](CODE_OF_CONDUCT.md).
+  By contributing to this project, you agree to abide by its terms.
