### 2019-11-02 4208193
#70 update code of conduct and site auto build
