+++ .travis.yml
@@ -11,3 +11,15 @@ r_packages:
 after_success:
   - Rscript -e 'library(covr); codecov()'
   - Rscript -e 'Sys.setenv(NOT_CRAN = "true"); goodpractice::gp()'
+  - Rscript -e 'library(pkgdown); build_site()'
+
+before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
+
+deploy:
+  provider: pages
+  skip-cleanup: true
+  github-token: $GITHUB_PAT
+  keep-history: true
+  local-dir: docs
+  on:
+    branch: master
