### 2019-11-02 c6dd69e
fix #70, fix download badge
