### 2019-11-02 717b312
fix #70, fix download badge
