### 2019-11-03 474e50a
Update CODE_OF_CONDUCT.md
