### 2019-11-03 388b499
fix #71, add citation, contribution guide and update cod of conduct
