+++ CONTRIBUTING.md
@@ -1,23 +1,86 @@
-:+1::tada: First off, thanks for taking the time to contribute! :tada::+1:
-## How to contribute to dbparser
-#### **Did you find a bug?**
+# Contributing to drugverse
 
-* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/Dainanahan/dbparser/issues).
+<!-- This CONTRIBUTING.md is adapted from https://gist.github.com/peterdesmet/e90a1b0dc17af6c12daf6e8b2f044e7c -->
 
-* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/Dainanahan/dbparser/issues/new). 
-Be sure to select the correct template (issue/enhancement request) and fill the required fields as much as necessary.
+First of all, thanks for considering contributing to drugverse! 👍 It's people like you that make it rewarding for us - the project maintainers - to work on drugverse. 😊
 
-#### **Do you have questions about the source code?**
+drugverse is an open source project, maintained by people who care. We are not directly funded to do so.
 
-* Ask any question about how to use dbparser by emailing me at (moh_fcis@yahoo.com).
+[repo]: https://github.com/Dainanahan/dbparser
+[issues]: https://github.com/Dainanahan/dbparser/issues
+[new_issue]: https://github.com/Dainanahan/dbparser/issues/new
+[website]: https://dainanahan.github.io/dbparser
+[citation]: https://dainanahan.github.io/dbparser/authors.html
+[email]: mohammed.ali@edu.dsti.institute
 
+## Code of conduct
 
-## Our Code of Conduct
-This project and everyone participating in it is governed by the [Dainanahan Code of Conduct](CODE_OF_CONDUCT.md).
-By participating, you are expected to uphold this code. Please report unacceptable behavior to [moh_fcis@yahoo.com](mailto:moh_fcis@yahoo.com)
+Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
 
-**dbparser** is a volunteer effort and part of drugVerse framework. We encourage you to pitch in and join the team ;)
+## How you can contribute
 
-Thanks! :heart: :heart: :heart:
+There are several ways you can contribute to this project. If you want to know more about why and how to contribute to open source projects like this one, see this [Open Source Guide](https://opensource.guide/how-to-contribute/).
 
-Dainanahan Team
+### Share the love ❤️
+
+Think **drugverse** is useful? Let others discover it, by telling them in person, via Twitter or a blog post.
+
+Using **drugverse** for a paper you are writing? Consider [citing it][citation].
+
+### Ask a question ⁉️
+
+Using our_package and got stuck? Browse the [documentation][website] to see if you can find a solution. Still stuck? Post your question as an [issue on GitHub][new_issue]. While we cannot offer user support, we'll try to do our best to address it, as questions often lead to better documentation or the discovery of bugs.
+
+Want to ask a question in private? Contact the package maintainer by [email][email].
+
+### Propose an idea 💡
+
+Have an idea for a new our_package feature? Take a look at the [documentation][website] and [issue list][issues] to see if it isn't included or suggested yet. If not, suggest your idea as an [issue on GitHub][new_issue]. While we can't promise to implement your idea, it helps to:
+
+* Explain in detail how it would work.
+* Keep the scope as narrow as possible.
+
+See below if you want to contribute code for your idea as well.
+
+### Report a bug 🐛
+
+Using our_package and discovered a bug? That's annoying! Don't let others have the same experience and report it as an [issue on GitHub][new_issue] so we can fix it. A good bug report makes it easier for us to do so, so please include:
+
[…]
