### 2019-11-08 #347 @noamross
Hello @MohammedFCIS, thank for your submission, I'll be your handling editor.

### Editor checks:

---

#### Editor comments

Here is output from `goodpractice::gp()`.  Test coverage is very good - reviewers should use `covr::report()` to inspect uncovered areas.  Use of `<-` or `=` is up the author, but should be consistent througout the package.  Consider line length on a case-by-case basis, but consider the *styler* package or something similar for code formatting.

The are all minor and can be addressed as I look for reviewers, which I am doing now!

```
── GP dbparser ─────────────────────────

It is good practice to

    R/drug_atc_codes_node_parser.R:74:NA
    R/drug_carrier_node_parser.R:26:NA
    R/drug_carrier_node_parser.R:27:NA
    R/drug_carrier_node_parser.R:28:NA
    R/drug_carrier_node_parser.R:29:NA
    ... and 186 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.
  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/drug_parsers.R:986:39
    R/drug_polypeptide_sub_node_parser.R:3:13

    R/dbparser.R:1:1
    R/dbparser.R:7:1
    R/dbparser.R:13:1
    R/drug_all_nodes_parser.R:14:1
    R/drug_all_nodes_parser.R:15:1
    ... and 587 more lines

    vapply() instead.

    tests/testthat/test_drug_all_nodes_parser.R:10:3

───────────────────────────────────────
---

Reviewers:
Due date:
