### 2019-11-09 51c0858
#66, avoid calling sapply as it is not type safe
