### 2019-11-09 0f4d1c7
#66, omit Date in DESCRIPTION to match good practice
