### 2019-11-09 974f2a2
#66, limit code lines to 80 character
