+++ R/drug_all_nodes_parser.R
@@ -11,8 +11,10 @@
 #'
 #' @param save_table boolean, save table in database if true.
 #' @param save_csv boolean, save csv version of parsed dataframe if true
-#' @param csv_path location to save csv files into it, default is current location, save_csv must be true
-#' @param override_csv override existing csv, if any, in case it is true in the new parse operation
+#' @param csv_path location to save csv files into it, default is current
+#' location, save_csv must be true
+#' @param override_csv override existing csv, if any, in case it is true in the
+#'  new parse operation
 #' @return all drug elements dataframes
 #'
 #' @examples
@@ -38,7 +40,8 @@
 #' # If the csv exist before read it and return its data.
 #' parse_drug_all(save_csv = TRUE, csv_path = TRUE)
 #'
-#' # save parsed dataframe as csv if it does not exist in current location and return parsed dataframe.
+#' # save parsed dataframe as csv if it does not exist in current location and
+#' return parsed dataframe.
 #' # If the csv exist override it and return it.
 #' parse_drug_all(save_csv = TRUE, csv_path = TRUE, override = TRUE)
 #' }
@@ -78,7 +81,8 @@ parse_drug_all <-
       parse_drug_categories(save_table, save_csv, csv_path, override_csv)
     message("Parsed Categories_Drug, 10/75")
     Affected_Organisms_Drug <-
-      parse_drug_affected_organisms(save_table, save_csv, csv_path, override_csv)
+      parse_drug_affected_organisms(save_table, save_csv, csv_path,
+                                    override_csv)
     message("Parsed Affected_Organisms_Drug, 11/75")
     Dosages_Drug <-
       parse_drug_dosages(save_table, save_csv, csv_path, override_csv)
@@ -99,10 +103,12 @@ parse_drug_all <-
       parse_drug_interactions(save_table, save_csv, csv_path, override_csv)
     message("Parsed Interactions_Drug, 17/75")
     Experimental_Properties_Drug <-
-      parse_drug_experimental_properties(save_table, save_csv, csv_path, override_csv)
+      parse_drug_experimental_properties(save_table, save_csv, csv_path,
+                                         override_csv)
     message("Parsed Experimental_Properties_Drug, 18/75")
     External_Identifiers_Drug <-
-      parse_drug_external_identifiers(save_table, save_csv, csv_path, override_csv)
+      parse_drug_external_identifiers(save_table, save_csv, csv_path,
+                                      override_csv)
     message("Parsed External_Identifiers_Drug, 19/75")
     External_Links_Drug <-
       parse_drug_external_links(save_table, save_csv, csv_path, override_csv)
@@ -111,7 +117,8 @@ parse_drug_all <-
       parse_drug_snp_effects(save_table, save_csv, csv_path, override_csv)
     message("Parsed SNP_Effects_Drug, 21/75")
     SNP_Adverse_Drug_Reactions_Drug <-
-      parse_drug_snp_adverse_drug_reactions(save_table, save_csv, csv_path, override_csv)
+      parse_drug_snp_adverse_drug_reactions(save_table, save_csv, csv_path,
+                                            override_csv)
     message("Parsed SNP_Adverse_Drug_Reactions_Drug, 22/75")
     ATC_Codes_Drug <-
       parse_drug_atc_codes(save_table, save_csv, csv_path, override_csv)
@@ -123,25 +130,32 @@ parse_drug_all <-
       parse_drug_carriers_articles(save_table, save_csv, csv_path, override_csv)
     message("Parsed Articles_Carrier_Drug, 25/75")
     Textbooks_Carrier_Drug <-
-      parse_drug_carriers_textbooks(save_table, save_csv, csv_path, override_csv)
+      parse_drug_carriers_textbooks(save_table, save_csv, csv_path,
[…]
