+++ R/drug_targets_node_parser.R
@@ -122,7 +122,7 @@ parse_drug_targets_polypeptides_external_identifiers <-
       drug_targets_polypeptide_external_identifiers <- readr::read_csv(path)
     } else {
       drug_targets_polypeptide_external_identifiers <-
-        map_df(pkg.env$children,
+        map_df(pkg_env$children,
                ~ get_targets_polypeptide_external_identifiers_df(.x)) %>%
         unique()
 
@@ -132,7 +132,7 @@ parse_drug_targets_polypeptides_external_identifiers <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
+        con = pkg_env$con,
         df = drug_targets_polypeptide_external_identifiers,
         table_name = "drug_targets_polypeptides_external_identifiers",
         save_table_only = TRUE
@@ -205,7 +205,7 @@ parse_drug_targets_polypeptides_synonyms <-
       drug_targets_polypeptide_synonyms <- readr::read_csv(path)
     } else {
       drug_targets_polypeptide_synonyms <-
-        map_df(pkg.env$children,
+        map_df(pkg_env$children,
                ~ get_targets_polypeptide_synonyms_df(.x)) %>%
         unique()
 
@@ -214,7 +214,7 @@ parse_drug_targets_polypeptides_synonyms <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
+        con = pkg_env$con,
         df = drug_targets_polypeptide_synonyms,
         table_name = "drug_targets_polypeptides_synonyms",
         save_table_only = TRUE
@@ -288,7 +288,7 @@ parse_drug_targets_polypeptides_pfams <-
       drug_targets_polypeptide_pfams <- readr::read_csv(path)
     } else {
       drug_targets_polypeptide_pfams <-
-        map_df(pkg.env$children,
+        map_df(pkg_env$children,
                ~ get_targets_polypeptide_pfams_df(.x)) %>%
         unique()
 
@@ -298,7 +298,7 @@ parse_drug_targets_polypeptides_pfams <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
+        con = pkg_env$con,
         df = drug_targets_polypeptide_pfams,
         table_name = "drug_targets_polypeptides_pfams",
         save_table_only = TRUE
@@ -374,7 +374,7 @@ parse_drug_targets_polypeptides_go_classifiers <-
       drug_targets_polypeptides_go_classifiers <- readr::read_csv(path)
     } else {
       drug_targets_polypeptides_go_classifiers <-
-        map_df(pkg.env$children,
+        map_df(pkg_env$children,
                ~ get_targets_polypeptide_go_classifiers_df(.x)) %>%
         unique()
 
@@ -384,7 +384,7 @@ parse_drug_targets_polypeptides_go_classifiers <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
+        con = pkg_env$con,
         df = drug_targets_polypeptides_go_classifiers,
         table_name = "drug_targets_polypeptides_go_classifiers",
         save_table_only = TRUE
@@ -452,7 +452,7 @@ parse_drug_targets_actions <- function(save_table = FALSE, save_csv = FALSE,
     drug_targets_actions <- readr::read_csv(path)
   } else {
     drug_targets_actions <-
-      map_df(pkg.env$children, ~ get_targets_actions_df(.x)) %>%
+      map_df(pkg_env$children, ~ get_targets_actions_df(.x)) %>%
       unique()
     write_csv(drug_targets_actions, save_csv, csv_path)
   }
[…]
