+++ R/drug_transporter_node_parser.R
@@ -122,7 +122,7 @@ parse_drug_transporters_actions <-
       drug_transporters_actions <- readr::read_csv(path)
     } else {
       drug_transporters_actions <-
-        map_df(pkg.env$children, ~ get_transporters_actions_df(.x)) %>%
+        map_df(pkg_env$children, ~ get_transporters_actions_df(.x)) %>%
         unique()
 
       write_csv(drug_transporters_actions, save_csv, csv_path)
@@ -135,7 +135,7 @@ parse_drug_transporters_actions <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
+        con = pkg_env$con,
         df = drug_transporters_actions,
         table_name = "drug_transporters_actions",
         save_table_only = TRUE
@@ -210,7 +210,7 @@ parse_drug_transporters_articles <-
       drug_transporters_articles <- readr::read_csv(path)
     } else {
       drug_transporters_articles <-
-        map_df(pkg.env$children, ~ get_transporters_articles_df(.x)) %>%
+        map_df(pkg_env$children, ~ get_transporters_articles_df(.x)) %>%
         unique()
 
       write_csv(drug_transporters_articles, save_csv, csv_path)
@@ -218,7 +218,7 @@ parse_drug_transporters_articles <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
+        con = pkg_env$con,
         df = drug_transporters_articles,
         table_name = "drug_transporters_articles",
         save_table_only = TRUE
@@ -294,7 +294,7 @@ parse_drug_transporters_textbooks <-
       drug_transporters_textbooks <- readr::read_csv(path)
     } else {
       drug_transporters_textbooks <-
-        map_df(pkg.env$children, ~ get_transporters_textbooks_df(.x)) %>%
+        map_df(pkg_env$children, ~ get_transporters_textbooks_df(.x)) %>%
         unique()
 
       write_csv(drug_transporters_textbooks, save_csv, csv_path)
@@ -302,7 +302,7 @@ parse_drug_transporters_textbooks <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
+        con = pkg_env$con,
         df = drug_transporters_textbooks,
         table_name = "drug_transporters_textbooks",
         save_table_only = TRUE
@@ -376,7 +376,7 @@ parse_drug_transporters_links <-
       drug_transporters_links <- readr::read_csv(path)
     } else {
       drug_transporters_links <-
-        map_df(pkg.env$children, ~ get_transporters_links_df(.x)) %>%
+        map_df(pkg_env$children, ~ get_transporters_links_df(.x)) %>%
         unique()
 
       write_csv(drug_transporters_links, save_csv, csv_path)
@@ -384,7 +384,7 @@ parse_drug_transporters_links <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
+        con = pkg_env$con,
         df = drug_transporters_links,
         table_name = "drug_transporters_links",
         save_table_only = TRUE
@@ -459,7 +459,7 @@ parse_drug_transporters_polypeptides <-
       drug_transporters_polypeptides <- readr::read_csv(path)
     } else {
       drug_transporters_polypeptides <-
-        map_df(pkg.env$children, ~ get_transporters_polypeptide_df(.x)) %>%
+        map_df(pkg_env$children, ~ get_transporters_polypeptide_df(.x)) %>%
         unique()
 
       write_csv(drug_transporters_polypeptides, save_csv, csv_path)
@@ -467,11 +467,11 @@ parse_drug_transporters_polypeptides <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg.env$con,
[…]
