### 2019-11-09 35fb5f6
#66, handle space issues and trim some long names
