### 2019-11-09 0575bef
#66, continue handling long names
