+++ R/drug_all_nodes_parser.R
@@ -145,18 +145,18 @@ parse_drug_all <-
                                                             save_csv, csv_path,
                                                             override_csv)
     message("parsed carr_poly_ext_identity, 29/75")
-    synonyms_polypeptide_carrier_drug <-
-      parse_drug_carriers_polypeptides_synonyms(save_table, save_csv, csv_path,
+    carr_polypeptides_syn <-
+      parse_carr_polypeptides_syn(save_table, save_csv, csv_path,
                                                 override_csv)
-    message("parsed synonyms_polypeptide_carrier_drug, 30/75")
-    pfams_polypeptide_carrier_drug <-
-      parse_drug_carriers_polypeptides_pfams(save_table, save_csv, csv_path,
+    message("parsed carr_polypeptides_syn, 30/75")
+    carr_polypeptides_pfams <-
+      parse_carr_polypeptides_pfams(save_table, save_csv, csv_path,
                                              override_csv)
-    message("parsed pfams_polypeptide_carrier_drug, 31/75")
-    go_classifiers_polypeptide_carrier_drug <-
-      parse_drug_carriers_polypeptides_go_classifiers(save_table, save_csv,
+    message("parsed carr_polypeptides_pfams, 31/75")
+    carr_polypeptides_go <-
+      parse_carr_polypeptides_go(save_table, save_csv,
                                                       csv_path, override_csv)
-    message("parsed go_classifiers_polypeptide_carrier_drug, 32/75")
+    message("parsed carr_polypeptides_go, 32/75")
     carriers_drug <-
       parse_drug_carriers(save_table, save_csv, csv_path, override_csv)
     message("parsed carriers_drug, 33/75")
@@ -179,23 +179,23 @@ parse_drug_all <-
       parse_drug_enzymes_polypeptides(save_table, save_csv, csv_path,
                                       override_csv)
     message("parsed polypeptides_enzyme_drug, 39/75")
-    external_identifiers_polypeptide_enzyme_drug <-
-      parse_drug_enzymes_polypeptides_external_identifiers(save_table,
+    enzy_poly_ext_identity <-
+      parse_enzy_poly_ext_identitys(save_table,
                                                            save_csv, csv_path,
                                                            override_csv)
     message("parsed external_identifiers_polypeptides_enzyme_drug, 40/75")
-    synonyms_polypeptides_enzyme_drug <-
-      parse_drug_enzymes_polypeptides_synonyms(save_table, save_csv, csv_path,
+    enzy_poly_syn<-
+      parse_enzy_poly_syn(save_table, save_csv, csv_path,
                                                override_csv)
-    message("parsed synonyms_polypeptides_enzyme_drug, 41/75")
+    message("parsed enzy_poly_syn, 41/75")
     pfams_polypeptides_enzyme_drug <-
       parse_drug_enzymes_polypeptides_pfams(save_table, save_csv, csv_path,
                                             override_csv)
     message("parsed pfams_polypeptides_enzyme_drug, 42/75")
-    go_classifiers_polypeptides_enzyme_drug <-
-      parse_drug_enzymes_polypeptides_go_classifiers(save_table, save_csv,
+    enzy_poly_go <-
+      parse_enzy_poly_go(save_table, save_csv,
                                                      csv_path, override_csv)
-    message("parsed go_classifiers_polypeptides_enzyme_drug, 43/75")
+    message("parsed enzy_poly_go, 43/75")
     enzymes_drug <-
       parse_drug_enzymes(save_table, save_csv, csv_path, override_csv)
     message("parsed enzyme_drug, 44/75")
@@ -223,23 +223,23 @@ parse_drug_all <-
     sequences_drug <-
[…]
