### 2019-11-09 2e0cc61
#66, continue handling long names
