+++ man/parse_snp_adverse_reactions.Rd
@@ -1,0 +1,68 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/drug_parsers.R
+\name{parse_snp_adverse_reactions}
+\alias{parse_snp_adverse_reactions}
+\title{Extracts the drug snp adverse drug reactions element and return data as
+tibble.}
+\usage{
+parse_snp_adverse_reactions(save_table = FALSE, save_csv = FALSE,
+  csv_path = ".", override_csv = FALSE)
+}
+\arguments{
+\item{save_table}{boolean, save table in database if true.}
+
+\item{save_csv}{boolean, save csv version of parsed dataframe if true}
+
+\item{csv_path}{location to save csv files into it, default is current
+location, save_csv must be true}
+
+\item{override_csv}{override existing csv, if any, in case it is true in the
+new parse operation}
+}
+\value{
+drug snp adverse drug reactions node attributes tibble
+}
+\description{
+\code{parse_snp_adverse_reactions} returns tibble of drug
+ snp adverse drug reactions elements.
+}
+\details{
+This functions extracts the groups element of drug node in drugbank
+xml database with the option to save it in a predefined database via
+\code{\link{open_db}} method. It takes one single optional argument to
+save the returned tibble in the database.
+It must be called after \code{\link{get_xml_db_rows}} function like
+any other parser function.
+If \code{\link{get_xml_db_rows}} is called before for any reason, so
+no need to call it again before calling this function.
+}
+\examples{
+\donttest{
+# return only the parsed dataframe
+parse_snp_adverse_reactions()
+
+# save in database and return parsed dataframe
+parse_snp_adverse_reactions(save_table = TRUE)
+
+# save parsed dataframe as csv if it does not exist in current location
+# and return parsed dataframe.
+# If the csv exist before read it and return its data.
+parse_snp_adverse_reactions(save_csv = TRUE)
+
+# save in database, save parsed dataframe as csv if it does not exist in
+# current location and return parsed dataframe.
+# If the csv exist before read it and return its data.
+parse_snp_adverse_reactions(ssave_table = TRUE, save_csv = TRUE)
+
+# save parsed dataframe as csv if it does not exist in given location
+# and return parsed dataframe.
+# If the csv exist before read it and return its data.
+parse_snp_adverse_reactions(save_csv = TRUE, csv_path = TRUE)
+
+# save parsed dataframe as csv if it does not exist in current location
+# and return parsed dataframe.
+# If the csv exist override it and return it.
+parse_snp_adverse_reactions(save_csv = TRUE, csv_path = TRUE,
+override = TRUE)
+}
+}
