+++ man/parse_trans_poly_go.Rd
@@ -1,0 +1,73 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/drug_transporter_node_parser.R
+\name{parse_trans_poly_go}
+\alias{parse_trans_poly_go}
+\title{Extracts the drug transporters polypeptides go
+classifiers element and return data as data frame.}
+\usage{
+parse_trans_poly_go(save_table = FALSE, save_csv = FALSE,
+  csv_path = ".", override_csv = FALSE)
+}
+\arguments{
+\item{save_table}{boolean, save table in database if true.}
+
+\item{save_csv}{boolean, save csv version of parsed dataframe if true}
+
+\item{csv_path}{location to save csv files into it, default is current
+location, save_csv must be true}
+
+\item{override_csv}{override existing csv, if any, in case it is true in the
+new parse operation}
+}
+\value{
+drug transporters polypeptides go classifiers node attributes
+date frame
+}
+\description{
+\code{parse_trans_poly_go}
+returns data frame of drug transporters polypeptides
+go classifiers elements.
+}
+\details{
+This functions extracts the transporters polypeptides go
+ classifiers element of drug node in drug bank
+xml database with the option to save it in a predefined database via
+\code{\link{open_db}} method. It takes one single optional argument to
+save the returned dataframe in the database.
+It must be called after \code{\link{get_xml_db_rows}} function like
+any other parser function.
+If \code{\link{get_xml_db_rows}} is called before for any reason, so
+no need to call it again before calling this function.
+}
+\examples{
+\donttest{
+# return only the parsed dataframe
+parse_trans_poly_go()
+
+# save in database and return parsed dataframe
+parse_trans_poly_go(save_table = TRUE)
+
+# save parsed dataframe as csv if it does not exist in current
+# location and return parsed dataframe.
+# If the csv exist before read it and return its data.
+parse_trans_poly_go(save_csv = TRUE)
+
+# save in database, save parsed dataframe as csv if it does not exist in
+#  current location and return parsed dataframe.
+# If the csv exist before read it and return its data.
+parse_trans_poly_go(ssave_table = TRUE,
+save_csv = TRUE)
+
+# save parsed dataframe as csv if it does not exist in given location
+# and return parsed dataframe.
+# If the csv exist before read it and return its data.
+parse_trans_poly_go(save_csv = TRUE,
+csv_path = TRUE)
+
+# save parsed dataframe as csv if it does not exist in current
+# location and return parsed dataframe.
+# If the csv exist override it and return it.
+parse_trans_poly_go(
+save_csv = TRUE, csv_path = TRUE, override = TRUE)
+}
+}
