### 2019-11-09 56af8ef
#66, fix check warnings
