+++ R/drug_all_nodes_parser.R
@@ -41,7 +41,7 @@
 #' parse_drug_all(save_csv = trUe, csv_path = trUe)
 #'
 #' # save parsed dataframe as csv if it does not exist in current location and
-#' return parsed dataframe.
+#' # return parsed dataframe.
 #' # if the csv exist override it and return it.
 #' parse_drug_all(save_csv = trUe, csv_path = trUe, override = trUe)
 #' }
