### 2019-11-10 56b425b
#66, cotinue code refactor to limit names length
