+++ R/drug_all_nodes_parser.R
@@ -65,9 +65,9 @@ parse_drug_all <-
     links_drug <-
       parse_drug_links(save_table, save_csv, csv_path, override_csv)
     message("parsed links_drug, 5/75")
-    synonyms_drug <-
-      parse_drug_synonyms(save_table, save_csv, csv_path, override_csv)
-    message("parsed synonyms_drug, 6/75")
+    syn_drug <-
+      parse_drug_syn(save_table, save_csv, csv_path, override_csv)
+    message("parsed syn_drug, 6/75")
     products_drug <-
       parse_drug_products(save_table, save_csv, csv_path, override_csv)
     message("parsed products_drug, 7/75")
@@ -81,8 +81,10 @@ parse_drug_all <-
       parse_drug_categories(save_table, save_csv, csv_path, override_csv)
     message("parsed categories_drug, 10/75")
     affected_organisms_drug <-
-      parse_drug_affected_organisms(save_table, save_csv, csv_path,
-                                    override_csv)
+      parse_drug_affected_organisms(
+        save_table, save_csv, csv_path,
+        override_csv
+      )
     message("parsed affected_organisms_drug, 11/75")
     dosages_drug <-
       parse_drug_dosages(save_table, save_csv, csv_path, override_csv)
@@ -103,12 +105,16 @@ parse_drug_all <-
       parse_drug_interactions(save_table, save_csv, csv_path, override_csv)
     message("parsed interactions_drug, 17/75")
     experimental_properties_drug <-
-      parse_drug_experimental_properties(save_table, save_csv, csv_path,
-                                         override_csv)
+      parse_drug_exp_prop(
+        save_table, save_csv, csv_path,
+        override_csv
+      )
     message("parsed experimental_properties_drug, 18/75")
     external_identifiers_drug <-
-      parse_drug_external_identifiers(save_table, save_csv, csv_path,
-                                      override_csv)
+      parse_drug_ex_identity(
+        save_table, save_csv, csv_path,
+        override_csv
+      )
     message("parsed external_identifiers_drug, 19/75")
     external_links_drug <-
       parse_drug_external_links(save_table, save_csv, csv_path, override_csv)
@@ -117,8 +123,10 @@ parse_drug_all <-
       parse_drug_snp_effects(save_table, save_csv, csv_path, override_csv)
     message("parsed snp_effects_drug, 21/75")
     snp_adverse_reactions <-
-      parse_snp_adverse_reactions(save_table, save_csv, csv_path,
-                                            override_csv)
+      parse_snp_adverse_reactions(
+        save_table, save_csv, csv_path,
+        override_csv
+      )
     message("parsed snp_adverse_reactions, 22/75")
     atc_codes_drug <-
       parse_drug_atc_codes(save_table, save_csv, csv_path, override_csv)
@@ -130,32 +138,44 @@ parse_drug_all <-
       parse_drug_carriers_articles(save_table, save_csv, csv_path, override_csv)
     message("parsed articles_carrier_drug, 25/75")
     textbooks_carrier_drug <-
-      parse_drug_carriers_textbooks(save_table, save_csv, csv_path,
-                                    override_csv)
+      parse_drug_carriers_textbooks(
+        save_table, save_csv, csv_path,
+        override_csv
+      )
     message("parsed textbooks_carrier_drug, 26/75")
     links_carrier_drug <-
       parse_drug_carriers_links(save_table, save_csv, csv_path, override_csv)
     message("parsed links_carrier_drug, 27/75")
     polypeptides_carrier_drug <-
-      parse_drug_carriers_polypeptides(save_table, save_csv, csv_path,
-                                       override_csv)
+      parse_carr_poly(
+        save_table, save_csv, csv_path,
+        override_csv
[…]
