+++ R/drug_parsers.R
@@ -443,11 +443,11 @@ parse_drug_links <-
   }
 
 
-#' Extracts the drug synonyms element and return data as tibble.
+#' Extracts the drug syn element and return data as tibble.
 #'
-#' \code{parse_drug_synonyms} returns tibble of drug synonyms elements.
+#' \code{parse_drug_syn} returns tibble of drug syn elements.
 #'
-#' This functions extracts the synonyms element of drug node in drugbank
+#' This functions extracts the syn element of drug node in drugbank
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. It takes one single optional argument to
 #' save the returned tibble in the database.
@@ -462,61 +462,61 @@ parse_drug_links <-
 #' location, save_csv must be true
 #' @param override_csv override existing csv, if any, in case it is true in the
 #'  new parse operation
-#' @return drug synonyms node attributes tibble
+#' @return drug syn node attributes tibble
 #'
 #' @examples
 #' \donttest{
 #' # return only the parsed dataframe
-#' parse_drug_synonyms()
+#' parse_drug_syn()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug_synonyms(save_table = TRUE)
+#' parse_drug_syn(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_synonyms(save_csv = TRUE)
+#' parse_drug_syn(save_csv = TRUE)
 #'
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_synonyms(ssave_table = TRUE, save_csv = TRUE)
+#' parse_drug_syn(ssave_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_synonyms(save_csv = TRUE, csv_path = TRUE)
+#' parse_drug_syn(save_csv = TRUE, csv_path = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist override it and return it.
-#' parse_drug_synonyms(save_csv = TRUE, csv_path = TRUE, override = TRUE)
+#' parse_drug_syn(save_csv = TRUE, csv_path = TRUE, override = TRUE)
 #' }
 #' @export
-parse_drug_synonyms <-
+parse_drug_syn <-
   function(save_table = FALSE,
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    path <- get_dataset_full_path("drug_synonyms", csv_path)
+    path <- get_dataset_full_path("drug_syn", csv_path)
     if (!override_csv & file.exists(path)) {
-      drug_synonyms <- readr::read_csv(path)
+      drug_syn <- readr::read_csv(path)
     } else {
-      drug_synonyms <-
-        map_df(pkg_env$children, ~ get_synonyms_df(.x)) %>% unique()
-      write_csv(drug_synonyms, save_csv, csv_path)
+      drug_syn <-
+        map_df(pkg_env$children, ~ get_syn_df(.x)) %>% unique()
+      write_csv(drug_syn, save_csv, csv_path)
     }
 
 
     if (save_table) {
       save_drug_sub(
         con = pkg_env$con,
-        df = drug_synonyms,
-        table_name = "drug_synonyms",
+        df = drug_syn,
+        table_name = "drug_syn",
[…]
