### 2019-11-10 579ddd7
#66, cotinue code refactor to limit names length
