### 2019-11-10 da0f52c
#66, resolve all name length issues
