+++ R/drug_transporter_node_parser.R
@@ -1,40 +1,40 @@
-get_transporters_df <- function(rec) {
+get_trans_df <- function(rec) {
   return(map_df(xmlChildren(rec[["transporters"]]),
                 ~ get_organizm_rec(.x, xmlValue(rec["drugbank-id"][[1]]))))
 }
 
-get_transporters_actions_df <- function(rec) {
+get_trans_actions_df <- function(rec) {
   return(map_df(xmlChildren(rec[["transporters"]]), ~ drug_sub_df(.x,
                                                                   "actions",
                                                                   id = "id")))
 }
 
-get_transporters_articles_df <- function(rec) {
+get_trans_articles_df <- function(rec) {
   return(map_df(
     xmlChildren(rec[["transporters"]]),
     ~ drug_sub_df(.x, "references", seconadary_node = "articles", id = "id")
   ))
 }
 
-get_transporters_textbooks_df <- function(rec) {
+get_trans_textbooks_df <- function(rec) {
   return(map_df(
     xmlChildren(rec[["transporters"]]),
     ~ drug_sub_df(.x, "references", seconadary_node = "textbooks", id = "id")
   ))
 }
 
-get_transporters_links_df <- function(rec) {
+get_trans_links_df <- function(rec) {
   return(map_df(
     xmlChildren(rec[["transporters"]]),
     ~ drug_sub_df(.x, "references", seconadary_node = "links", id = "id")
   ))
 }
 
-get_transporters_polypeptide_df <- function(rec) {
+get_trans_poly_df <- function(rec) {
   return(map_df(xmlChildren(rec[["transporters"]]), ~ get_polypeptide_rec(.x)))
 }
 
-get_transporters_poly_ex_identity_df <-
+get_trans_poly_ex_identity_df <-
   function(rec) {
     return(map_df(
       xmlChildren(rec[["transporters"]]),
@@ -42,17 +42,17 @@ get_transporters_poly_ex_identity_df <-
     ))
   }
 
-get_transporters_polypeptide_syn_df <- function(rec) {
+get_trans_poly_syn_df <- function(rec) {
   return(map_df(xmlChildren(rec[["transporters"]]),
                 ~ get_polypeptide_syn(.x)))
 }
 
-get_transporters_polypeptide_pfams_df <- function(rec) {
+get_trans_poly_pfams_df <- function(rec) {
   return(map_df(xmlChildren(rec[["transporters"]]),
                 ~ get_polypeptide_pfams(.x)))
 }
 
-get_transporters_polypeptide_go_df <- function(rec) {
+get_trans_poly_go_df <- function(rec) {
   return(map_df(xmlChildren(rec[["transporters"]]),
                 ~ get_polypeptide_go(.x)))
 }
@@ -84,64 +84,64 @@ get_transporters_polypeptide_go_df <- function(rec) {
 #' @examples
 #' \donttest{
 #' # return only the parsed dataframe
-#' parse_drug_transporters_actions()
+#' parse_drug_trans_actions()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug_transporters_actions(save_table = TRUE)
+#' parse_drug_trans_actions(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current location
 #' # and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_transporters_actions(save_csv = TRUE)
+#' parse_drug_trans_actions(save_csv = TRUE)
 #'
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
[…]
