### 2019-11-10 feea346
#66, resolve global bane binding issue
