+++ R/drug_common_utilities.R
@@ -79,9 +79,9 @@ get_xml_db_rows <- function(xml_db_name) {
 
   if (ext == "zip") {
     tryCatch({
-      db <- unzip(xml_db_name, exdir = dir_name)
+      unzip(xml_db_name, exdir = dir_name)
       db <- unzip(xml_db_name, list = TRUE)
-      xml_db_name <- paste0(dir_name, "/", db$Name[1])
+      xml_db_name <- paste0(dir_name, "/", db[[1]])
       message(xml_db_name)
     },
     error = function(e) {
