+++ tests/testthat/test_drug_carrier_node_parser_biotech.R
@@ -7,76 +7,114 @@ library(tibble)
 library(purrr)
 
 biotech <- "drugbank_record_biotech.xml"
-test_that(desc = "Read database",
-          code = {
-            expect_true(get_xml_db_rows(
-              system.file("extdata", biotech, package = "dbparser")
-            ))
-          })
+test_that(
+  desc = "Read database",
+  code = {
+    expect_true(get_xml_db_rows(
+      system.file("extdata", biotech, package = "dbparser")
+    ))
+  }
+)
 
-test_that(desc = "Read drug carriers actions that",
-          code = {
-            expect_equal(nrow(parse_drug_carriers_actions()),
-                         0)
-            expect_error(parse_drug_carriers_actions(TRUE))
-          })
+test_that(
+  desc = "Read drug carriers actions that",
+  code = {
+    expect_equal(
+      nrow(parse_drug_carriers_actions()),
+      0
+    )
+    expect_error(parse_drug_carriers_actions(TRUE))
+  }
+)
 
-test_that(desc = "Read drug carriers articles that",
-          code = {
-            expect_equal(nrow(parse_drug_carriers_articles()),
-                         0)
-            expect_error(parse_drug_carriers_articles(TRUE))
-          })
+test_that(
+  desc = "Read drug carriers articles that",
+  code = {
+    expect_equal(
+      nrow(parse_drug_carriers_articles()),
+      0
+    )
+    expect_error(parse_drug_carriers_articles(TRUE))
+  }
+)
 
-test_that(desc = "Read drug carriers_textbooks that",
-          code = {
-            expect_equal(nrow(parse_drug_carriers_textbooks()),
-                         0)
-            expect_error(parse_drug_carriers_textbooks(TRUE))
-          })
+test_that(
+  desc = "Read drug carriers_textbooks that",
+  code = {
+    expect_equal(
+      nrow(parse_drug_carriers_textbooks()),
+      0
+    )
+    expect_error(parse_drug_carriers_textbooks(TRUE))
+  }
+)
 
-test_that(desc = "Read drug carriers polypeptides that",
-          code = {
-            expect_equal(nrow(parse_carr_poly()),
-                         0)
-            expect_error(parse_carr_poly(TRUE))
-          })
+test_that(
+  desc = "Read drug carriers polypeptides that",
+  code = {
+    expect_equal(
+      nrow(parse_carr_poly()),
+      0
+    )
+    expect_error(parse_carr_poly(TRUE))
+  }
+)
 
-test_that(desc = "Read drug carriers polypeptides external identifiers that",
-          code = {
-            expect_equal(nrow(
-              parse_carr_poly_ext_identity()
-            ),
-            0)
-            expect_error(
-              parse_carr_poly_ext_identity(TRUE))
-          })
+test_that(
+  desc = "Read drug carriers polypeptides external identifiers that",
+  code = {
+    expect_equal(
+      nrow(
+        parse_carr_poly_ext_identity()
+      ),
+      0
+    )
+    expect_error(
+      parse_carr_poly_ext_identity(TRUE)
+    )
+  }
+)
 
-test_that(desc = "Read drug carriers polypeptides syn that",
-          code = {
-            expect_equal(nrow(parse_carr_polypeptides_syn()),
-                         0)
-            expect_error(parse_carr_polypeptides_syn(TRUE))
-          })
+test_that(
+  desc = "Read drug carriers polypeptides syn that",
+  code = {
+    expect_equal(
+      nrow(parse_carr_polypeptides_syn()),
+      0
+    )
+    expect_error(parse_carr_polypeptides_syn(TRUE))
+  }
+)
 
[…]
