### 2019-11-10 #347 @MohammedFCIS
Hello @noamross , I have fixed all goodpractice notes, except test coverage that increased to 92%, is it necessary to increase the test coverage to 100% before resuming review? Thank you
