### 2019-11-18 #347 @noamross
Thank you @MohammedFCIS.  No, 100% test coverage is not required.

Thank you @emmamendelsohn and @haozhu233 for agreeing to review!  Reviews are due 2019-12-09.
