### 2019-11-18 b4d4aa4
add ropensci badge
