+++ README.md
@@ -19,6 +19,7 @@ downloads](https://cranlogs.r-pkg.org/badges/grand-total/dbparser)](https://cran
 [![Rdoc](http://www.rdocumentation.org/badges/version/dbparser)](http://www.rdocumentation.org/packages/dbparser)
 [![CII Best
 Practices](https://bestpractices.coreinfrastructure.org/projects/3311/badge)](https://bestpractices.coreinfrastructure.org/projects/3311)
+[![](https://badges.ropensci.org/347_status.svg)](https://github.com/ropensci/software-review/issues/347)
 
 The main purpose of the `dbparser` package is to parse the
 [DrugBank](http://drugbank.ca/) database which is downloadable in XML
