### 2019-12-02 #347 @emmamendelsohn
Hi @MohammedFCIS! I'm in the process of acquiring the DrugBank dataset--I just responded to a request for more information about how I will use the data. Hopefully they will approve me soon! In the meantime, I was wondering if you considered using the [DrugBank API](https://docs.drugbankplus.com/v1/#introduction)? Could it potentially provide a smoother path for users to access the data?
