### 2019-12-04 #347 @MohammedFCIS
Hi @emmamendelsohn , first let me thank you so much for taking time and effort to review our package.
As for the DrugBank database I have the following backup in my google drive and you can use it if you like :)
https://drive.google.com/open?id=1Em-IYnuHeOENE9qoiZxX2QGr9xE2fS_r.

Regarding DrugBank API, we considered using it at first of course, but decided to proceed with our dbparser for the following reasons:
- We needed DrugBank database parser to parse any piece of information we want, not just those available on the DrugBank API.
- We needed to use these information as R tibbles for developed drugs development machine learning algorithms.
- Moreover, we decided to build a complete ecosystem system for drugs development https://dainanahan.github.io/drugverse/, and dbparser is just the begining ;).
- We wanted to make a contribution to open science community as general.
- I am not sure if those API's are free as well :)

I hope that answered your question :). Thank you.
