### 2019-12-09 #347 @haozhu233
Hi, all,

Here is my review. @MohammedFCIS Good job on this package! I totally enjoyed reviewing it! :)

Hao

--------------

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: 4 hr

---

The `Drugbank` database provides quite some useful information for the bioinformatics and cheminformatics research. However, for new researchers, especially for those who are not familiar with the xml format, it might be a little difficult for them to get their hands on the data. The `dbparser` package provides a set of functions to parse the drugbank xml data file. I think as a data extraction tool, it achieves its purpose and can bring a lot of convenience to its users. The code in this package are well written and documented with a few exceptions, which is listed below.

#### Some notes and observations:
1. I noticed that right now DrugBank is released as version 5 (in fact 5.1.4 as of today). I would imagine there could be inconsistencies in the database between each major release. I wonder if this package can be used to read data from DrugBank version 4.0 or 6.0 without getting an error. Have you checked what kind of changes they made for every major release? If this package only works with version 5, it might be a good idea to note it in DESCRIPTION or README, or add a database version check in `get_xml_db_rows`.

2. rOpenSci strongly recommend `xml2` for parsing xml data whereas this package uses `XML`. I guess at this point it might be a little too late and overkill to ask you to switch to `xml2` but hopefully you could consider to use that next time. :)

See https://devguide.ropensci.org/building.html#recommended-scaffolding

3. I noticed that you tried to fix some function names that are too long a few weeks ago. It's generally a good idea to keep variable names short but I think your fixes brought up some inconsistencies among functions under the same umbrella. For example, all the following functions work with drug carrier data. Why do some functions start with `parse_carr` while others start with `parse_drug_carriers`? In fact, I think users of this package already know that they are tying to parse something. Do they really need to type `parse` every time to get some data? How about replacing it with something like `db_` (or something more appropriate).

See https://devguide.ropensci.org/building.html#function-and-argument-naming

4. `get_xml_db_rows` seems to be the name of a secondary function but is in fact the initial step of reading in the data.

5. Should this file existence check take place at the very beginning?
https://github.com/Dainanahan/dbparser/blob/b4d4aa47dab24e4eb291eec96ba5f7bbeb4275ef/R/drug_common_utilities.R#L96

6. I think the "Exploring the data" section is a little too heavy for README. This kind of contents is great in the getting started vignette but not that telling in README as it's just using dplyr and ggplot.

Package check passed without warning/notes on my machine.
