### 2019-12-10 #347 @emmamendelsohn
@MohammedFCIS Nice job and thanks for the opportunity to review. Happy to discuss/answer questions.

---

`dbparser` is a convenient package for parsing XML DrugBank data into tidy format. I have a few minor comments on documentation, and then a more general question about saving the data into a database.

Tests run fine with `devtools::test()`. Running testthat.R yields an error: `Error: No tests found for dbparser`.

#### notes on README
- I think it would help unfamiliar users to add another sentence in the introduction explaining what DrugBank is. It would also be helpful to include a note about the need to make an account and request access to the data.
- Suggest removing commented code
- It seems like you made some changes recently that are resulting in errors in the README, so I suggest combing through it to make sure it's working as expected. Below are a couple things I've noticed.
  - `parse_drug_targets_actions()` no longer works (it's now `parse_drug_targ_actions()`). Side note: I agree with @haozhu233 it would be nice to shorten some of the function names if possible.
  - `xml_db_name` is not an argument in `open_db()`
 - As expected `parse_drug(save_table = TRUE)` (and all parse functions) throws an error when there is no database to save to. However, the error only happens after parsing the data for about a minute, so it would be nice to throw an early informative error to the user.
 - Make sure all relevant libraries are called for the plots (`ggplot2`, `dplyr`)
 - The package description in R help can use a bit of clarification. I'd suggest emphasizing the drug bank parser over the database functionality.

#### Database
This may be due to inexperience on my part, but I was confused about how to save the database using the driver and server argument. I would have liked to be able to save it to a SQLite or other local database using `DBI::dbConnect`. Is it possible to make these functions more generalizable? Or at minimum, add more description of what type of databases the package supports?
