### 2019-12-10 Dainanahan/dbparser#73 @MohammedFCIS
Track DrugBank version that package support in Readme
Note from ropensci

> I noticed that right now DrugBank is released under version 5 (5.1.4 as of today). I would imagine there could be inconsistencies in the database between each major release. I wonder if this package can be used to read data from DrugBank version 4.0 or 6.0 without getting an error. Have you checked what kind of changes they made for every major release? If this package only works with version 5, it might be a good idea to note it in DESCRIPTION or README, or add a database version check in get_xml_db_rows.
