### 2019-12-10 Dainanahan/dbparser#74 @MohammedFCIS
Make function names consistence 
A note from ropensci

> I noticed that you tried to fix some function names that are too long a few weeks ago. It's generally a good idea to keep variable names short but I think your fixes brought up some inconsistencies among functions under the same umbrella. For example, all the functions in the screenshot below work with drug carrier data. Why do some functions start with parse_carr while others start with parse_drug_carriers? In fact, I think users of this package already know that they are tying to parse something. Do they really need to type parse_ or parse_drug_ every time when they want to get some data? How about replacing it with something like db_ (or something more appropriate)
