### 2019-12-10 Dainanahan/dbparser#75 @MohammedFCIS
Change get_xml_db_rows name to be more meaningful 
A note from ropensci
>get_xml_db_rows seems to be the name of a secondary function but is in fact the initial step of reading in the data.
