### 2019-12-10 #347 @MohammedFCIS
Hello @haozhu233  and @emmamendelsohn ;
Thank you so much for your time and effort that you are giving for this review.
I did not only benefited from your comments but I enjoyed your feedback so much as well, thanks for that also.
Kindly find my detailed response below:

> I noticed that right now DrugBank is released under version 5 (5.1.4 as of today). I would imagine there could be inconsistencies in the database between each major release. I wonder if this package can be used to read data from DrugBank version 4.0 or 6.0 without getting an error. Have you checked what kind of changes they made for every major release? If this package only works with version 5, it might be a good idea to note it in DESCRIPTION or README, or add a database version check in get_xml_db_rows.

I agree with you, the following issue has been created for that.
https://github.com/Dainanahan/dbparser/issues/73

> rOpenSci strongly recommend xml2 for parsing xml data whereas this package uses XML. I guess at this point it might be a little too late and overkill to ask you to switch to xml2 but hopefully you could consider to use that next time. :)

Actually it is in my plan to migrate to ```xml2```, kindly check that issue:
https://github.com/Dainanahan/dbparser/issues/63

> I noticed that you tried to fix some function names that are too long a few weeks ago. It's generally a good idea to keep variable names short but I think your fixes brought up some inconsistencies among functions under the same umbrella. For example, all the functions in the screenshot below work with drug carrier data. Why do some functions start with parse_carr while others start with parse_drug_carriers? In fact, I think users of this package already know that they are tying to parse something. Do they really need to type parse_ or parse_drug_ every time when they want to get some data? How about replacing it with something like db_ (or something more appropriate)

The following issue is created to fix that https://github.com/Dainanahan/dbparser/issues/74

> get_xml_db_rows seems to be the name of a secondary function but is in fact the initial step of reading in the data.

The following issue is created to fix that https://github.com/Dainanahan/dbparser/issues/75

> Should this file existence check take place at the very beginning?
https://github.com/Dainanahan/dbparser/blob/b4d4aa47dab24e4eb291eec96ba5f7bbeb4275ef/R/drug_common_utilities.R#L96

It is in the beginning already, nothing before it but **zip** file check. As if the file we have is a **zip** file we must unzip before checking for the db file.

> I think the "Exploring the data" section is a little too heavy for README. This kind of contents is great in the getting started vignette but is not that useful in README as it is basically demonstrating dplyr and ggplot.

Added this issue to fix it https://github.com/Dainanahan/dbparser/issues/14
that to cover @emmamendelsohn  notes as well.
I will start working on your notes right away.
Best Regards,
