+++ R/drug_classfication_node_parser.R
@@ -15,7 +15,7 @@ drug_classifications_df <- function(rec) {
 #' \code{parse_drug_classification} returns data frame of drug classifications
 #'  elements.
 #'
-#' This functions extracts the classifications element of drug node in drug bank
+#' This functions extracts the classifications element of drug node in \strong{DrugBank}
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. It takes one single optional argument to
 #' save the returned dataframe in the database.
