+++ R/drug_price_node_parser.R
@@ -20,7 +20,7 @@ get_prices_df <- function(rec) {
 #'
 #' \code{parse_drug_prices} returns data frame of drug prices elements.
 #'
-#' This functions extracts the prices element of drug node in drug bank
+#' This functions extracts the prices element of drug node in \strong{DrugBank}
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. It takes one single optional argument to
 #' save the returned dataframe in the database.

[…]
