+++ man/parse_drug_targ.Rd
@@ -25,7 +25,7 @@ drug target node attributes date frame
 \code{parse_drug_targ} returns data frame of drug targ elements.
 }
 \details{
-This functions extracts the target element of drug node in drug bank
+This functions extracts the target element of drug node in \strong{DrugBank}
 xml database with the option to save it in a predefined database via
 \code{\link{open_db}} method. It takes one single optional argument to
 save the returned dataframe in the database.
