+++ man/parse_drug_element.Rd
@@ -29,7 +29,7 @@ list of selected drug elements dataframes
 elements.
 }
 \details{
-this functions extracts selected element of drug nodes in drug bank
+this functions extracts selected element of drug nodes in \strong{DrugBank}
 xml database with the option to save it in a predefined database via
 \code{\link{open_db}} method. it takes one single optional argument to
 save the returned dataframe in the database.
