+++ man/parse_drug_classification.Rd
@@ -26,7 +26,7 @@ drug classifications node attributes date frame
  elements.
 }
 \details{
-This functions extracts the classifications element of drug node in drug bank
+This functions extracts the classifications element of drug node in \strong{DrugBank}
 xml database with the option to save it in a predefined database via
 \code{\link{open_db}} method. It takes one single optional argument to
 save the returned dataframe in the database.
