+++ man/parse_carr_poly_ext_identity.Rd
@@ -29,7 +29,7 @@ drug carriers polypeptides external identifiers
 }
 \details{
 This functions extracts the carriers polypeptides external identifiers
- element of drug node in drug bank
+ element of drug node in \strong{DrugBank}
 xml database with the option to save it in a predefined database via
 \code{\link{open_db}} method. It takes one single optional argument to
 save the returned dataframe in the database.
@@ -41,32 +41,37 @@ no need to call it again before calling this function.
 \examples{
 \donttest{
 # return only the parsed dataframe
-parse_carr_poly_ext_identity ()
+parse_carr_poly_ext_identity()
 
 # save in database and return parsed dataframe
-parse_carr_poly_ext_identity (save_table = TRUE)
+parse_carr_poly_ext_identity(save_table = TRUE)
 
 # save parsed dataframe as csv if it does not exist in current
 # location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_carr_poly_ext_identity (save_csv = TRUE)
+parse_carr_poly_ext_identity(save_csv = TRUE)
 
 # save in database, save parsed dataframe as csv if it does not exist
 # in current location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_carr_poly_ext_identity (ssave_table = TRUE,
- save_csv = TRUE)
+parse_carr_poly_ext_identity(
+  ssave_table = TRUE,
+  save_csv = TRUE
+)
 
 # save parsed dataframe as csv if it does not exist in given
 # location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_carr_poly_ext_identity (save_csv = TRUE,
-csv_path = TRUE)
+parse_carr_poly_ext_identity(
+  save_csv = TRUE,
+  csv_path = TRUE
+)
 
 # save parsed dataframe as csv if it does not exist in current
 # location and return parsed dataframe.
 # If the csv exist override it and return it.
-parse_carr_poly_ext_identity (
-save_csv = TRUE, csv_path = TRUE, override = TRUE)
+parse_carr_poly_ext_identity(
+  save_csv = TRUE, csv_path = TRUE, override = TRUE
+)
 }
 }
