+++ man/parse_drug_carriers_articles.Rd
@@ -63,7 +63,9 @@ parse_drug_carriers_articles(save_csv = TRUE, csv_path = TRUE)
 # save parsed dataframe as csv if it does not exist in current
 # location and return parsed dataframe.
 # If the csv exist override it and return it.
-parse_drug_carriers_articles(save_csv = TRUE, csv_path = TRUE,
-override = TRUE)
+parse_drug_carriers_articles(
+  save_csv = TRUE, csv_path = TRUE,
+  override = TRUE
+)
 }
 }
