### 2019-12-11 4d821a3
#75, change DrugBank db reader function name
