### 2019-12-11 3c5efc9
#14, use donotrun instead of donottest
