+++ R/drug_price_node_parser.R
@@ -38,7 +38,7 @@ get_prices_df <- function(rec) {
 #' @return drug prices node attributes date frame
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' # return only the parsed dataframe
 #' parse_drug_prices()
 #'
