### 2019-12-11 8b65595
#75, replace old drugbank function reader in documentation
