+++ R/drug_classfication_node_parser.R
@@ -19,9 +19,9 @@ drug_classifications_df <- function(rec) {
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. It takes one single optional argument to
 #' save the returned dataframe in the database.
-#' It must be called after \code{\link{get_xml_db_rows}} function like
+#' It must be called after \code{\link{read_drugbank_xml_db}} function like
 #' any other parser function.
-#' If \code{\link{get_xml_db_rows}} is called before for any reason, so
+#' If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
 #' no need to call it again before calling this function.
 #'
 #' @param save_table boolean, save table in database if true.
