+++ tests/testthat/test_drug_main_node_parser_small_molecule.R
@@ -11,7 +11,7 @@ small_molecule <- "drugbank_record_small_molecule.xml"
 test_that(
   desc = "Read database",
   code = {
-    expect_true(get_xml_db_rows(
+    expect_true(read_drugbank_xml_db(
       system.file("extdata", small_molecule, package = "dbparser")
     ))
   }
