+++ R/drug_all_nodes_parser.R
@@ -4,9 +4,9 @@
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. it takes one single optional argument to
 #' save the returned dataframe in the database.
-#' it must be called after \code{\link{get_xml_db_rows}} function like
+#' it must be called after \code{\link{read_drugbank_xml_db}} function like
 #' any other parser function.
-#' if \code{\link{get_xml_db_rows}} is called before for any reason, so
+#' if \code{\link{read_drugbank_xml_db}} is called before for any reason, so
 #' no need to call it again before calling this function.
 #'
 #' @param save_table boolean, save table in database if true.
@@ -472,9 +472,9 @@ parse_drug_all <-
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. it takes one single optional argument to
 #' save the returned dataframe in the database.
-#' it must be called after \code{\link{get_xml_db_rows}} function like
+#' it must be called after \code{\link{read_drugbank_xml_db}} function like
 #' any other parser function.
-#' if \code{\link{get_xml_db_rows}} is called before for any reason, so
+#' if \code{\link{read_drugbank_xml_db}} is called before for any reason, so
 #' no need to call it again before calling this function.
 #'
 #' parse_drug_element_options can be called to know the valid options for
