+++ tests/testthat/test_drug_transporter_node_parser_small_molecule.R
@@ -10,7 +10,7 @@ small_molecule <- "drugbank_record_small_molecule.xml"
 test_that(
   desc = "Read database",
   code = {
-    expect_true(get_xml_db_rows(
+    expect_true(read_drugbank_xml_db(
       system.file("extdata", small_molecule, package = "dbparser")
     ))
   }
