+++ tests/testthat/test_drug_enzyme_node_parser_biotech.R
@@ -11,7 +11,7 @@ test_that(
   desc = "Read database",
   code = {
     expect_true(
-      get_xml_db_rows(
+      read_drugbank_xml_db(
         system.file("extdata", biotech, package = "dbparser")
       )
     )
