### 2019-12-11 19d6987
fix #75, update function documentation
