### 2019-12-12 #347 @noamross
Thank you for your reviews, @haozhu233 and @emmamendelsohn.  Thank you for responding quickly @MohammedFCIS and we look forward to the revision when you have addressed all the comments.  I concur with @emmamendelsohn's comment about the database - I'm not sure if there is a specific reason that MySQL is used, but I think it would make more sense to just let the user use any back-end database they pass.  I think you're already doing this.
