### 2019-12-23 1b8e643
$74, update database methods reference
