+++ tests/testthat/test_drug_carrier_node_parser_biotech.R
@@ -20,10 +20,10 @@ test_that(
   desc = "Read drug carriers actions that",
   code = {
     expect_equal(
-      nrow(parse_drug_carriers_actions()),
+      nrow(carriers_actions()),
       0
     )
-    expect_error(parse_drug_carriers_actions(TRUE))
+    expect_error(carriers_actions(TRUE))
   }
 )
 
@@ -31,10 +31,10 @@ test_that(
   desc = "Read drug carriers articles that",
   code = {
     expect_equal(
-      nrow(parse_drug_carriers_articles()),
+      nrow(carriers_articles()),
       0
     )
-    expect_error(parse_drug_carriers_articles(TRUE))
+    expect_error(carriers_articles(TRUE))
   }
 )
 
@@ -42,10 +42,10 @@ test_that(
   desc = "Read drug carriers_textbooks that",
   code = {
     expect_equal(
-      nrow(parse_drug_carriers_textbooks()),
+      nrow(carriers_textbooks()),
       0
     )
-    expect_error(parse_drug_carriers_textbooks(TRUE))
+    expect_error(carriers_textbooks(TRUE))
   }
 )
 
@@ -53,10 +53,10 @@ test_that(
   desc = "Read drug carriers polypeptides that",
   code = {
     expect_equal(
-      nrow(parse_carr_poly()),
+      nrow(carriers_polypeptide()),
       0
     )
-    expect_error(parse_carr_poly(TRUE))
+    expect_error(carriers_polypeptide(TRUE))
   }
 )
 
@@ -65,12 +65,12 @@ test_that(
   code = {
     expect_equal(
       nrow(
-        parse_carr_poly_ext_identity()
+        carriers_polypeptide_ext_identity()
       ),
       0
     )
     expect_error(
-      parse_carr_poly_ext_identity(TRUE)
+      carriers_polypeptide_ext_identity(TRUE)
     )
   }
 )
@@ -79,10 +79,10 @@ test_that(
   desc = "Read drug carriers polypeptides syn that",
   code = {
     expect_equal(
-      nrow(parse_carr_polypeptides_syn()),
+      nrow(carriers_polypeptidepeptides_syn()),
       0
     )
-    expect_error(parse_carr_polypeptides_syn(TRUE))
+    expect_error(carriers_polypeptidepeptides_syn(TRUE))
   }
 )
 
@@ -90,10 +90,10 @@ test_that(
   desc = "Read drug carriers polypeptides go classifiers that",
   code = {
     expect_equal(
-      nrow(parse_carr_polypeptides_go()),
+      nrow(carriers_polypeptidepeptides_go()),
       0
     )
-    expect_error(parse_carr_polypeptides_go(TRUE))
+    expect_error(carriers_polypeptidepeptides_go(TRUE))
   }
 )
 
@@ -112,9 +112,9 @@ test_that(
   desc = "Read drug carriers polypeptides pfams that",
   code = {
     expect_equal(
-      nrow(parse_carr_polypeptides_pfams()),
+      nrow(carriers_polypeptidepeptides_pfams()),
       0
     )
-    expect_error(parse_carr_polypeptides_pfams(TRUE))
+    expect_error(carriers_polypeptidepeptides_pfams(TRUE))
   }
 )
