+++ man/carriers_textbooks.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/drug_carrier_node_parser.R
-\name{parse_drug_carriers_links}
-\alias{parse_drug_carriers_links}
-\title{Extracts the drug carriers links element and return data as data frame.}
+\name{carriers_textbooks}
+\alias{carriers_textbooks}
+\title{Extracts the drug carriers textbooks element and return data as data frame.}
 \usage{
-parse_drug_carriers_links(save_table = FALSE, save_csv = FALSE,
+carriers_textbooks(save_table = FALSE, save_csv = FALSE,
   csv_path = ".", override_csv = FALSE)
 }
 \arguments{
@@ -15,17 +15,19 @@ parse_drug_carriers_links(save_table = FALSE, save_csv = FALSE,
 \item{csv_path}{location to save csv files into it, default is current
 location, save_csv must be true}
 
-\item{override_csv}{override existing csv, if any, in case it is true
-in the new parse operation}
+\item{override_csv}{override existing csv, if any, in case it is true in the
+new parse operation}
 }
 \value{
-drug carriers_links node attributes date frame
+drug carriers textbooks node attributes date frame
 }
 \description{
-\code{parse_drug_groups} returns data frame of drug carriers links elements.
+\code{carriers_textbooks} returns data frame of drug carriers
+ textbooks elements.
 }
 \details{
-This functions extracts the carriers links element of drug node in \strong{DrugBank}
+This functions extracts the carriers textbooks element of drug node in
+drugbank
 xml database with the option to save it in a predefined database via
 \code{\link{open_db}} method. It takes one single optional argument to
 save the returned dataframe in the database.
@@ -37,30 +39,44 @@ no need to call it again before calling this function.
 \examples{
 \dontrun{
 # return only the parsed dataframe
-parse_drug_carriers_links()
+carriers_textbooks()
 
 # save in database and return parsed dataframe
-parse_drug_carriers_links(save_table = TRUE)
+carriers_textbooks(save_table = TRUE)
 
 # save parsed dataframe as csv if it does not exist in current location
 # and return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_drug_carriers_links(save_csv = TRUE)
+carriers_textbooks(save_csv = TRUE)
 
 # save in database, save parsed dataframe as csv if it does not exist in
-current
-# location and return parsed dataframe.
+# current location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_drug_carriers_links(ssave_table = TRUE, save_csv = TRUE)
+carriers_textbooks(ssave_table = TRUE, save_csv = TRUE)
 
 # save parsed dataframe as csv if it does not exist in given location and
 # return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_drug_carriers_links(save_csv = TRUE, csv_path = TRUE)
+carriers_textbooks(save_csv = TRUE, csv_path = TRUE)
 
 # save parsed dataframe as csv if it does not exist in current location and
 # return parsed dataframe.
 # If the csv exist override it and return it.
-parse_drug_carriers_links(save_csv = TRUE, csv_path = TRUE, override = TRUE)
+carriers_textbooks(
+  save_csv = TRUE, csv_path = TRUE, override =
+    TRUE
+)
 }
 }
+\seealso{
+Other carriers: \code{\link{carriers_actions}},
+  \code{\link{carriers_articles}},
+  \code{\link{carriers_links}},
+  \code{\link{carriers_polypeptide_ext_identity}},
[…]

[…]
