### 2019-12-23 1463593
#74, continue to update carrier functions
