+++ tests/testthat/test_drug_carrier_node_parser_small_molecule.R
@@ -99,10 +99,10 @@ test_that(
   desc = "Read drug carriers that",
   code = {
     expect_equal(
-      nrow(parse_drug_carriers()),
+      nrow(carriers()),
       0
     )
-    expect_error(parse_drug_carriers(TRUE))
+    expect_error(carriers(TRUE))
   }
 )
 
