+++ man/carriers.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/drug_carrier_node_parser.R
-\name{parse_drug_carriers}
-\alias{parse_drug_carriers}
+\name{carriers}
+\alias{carriers}
 \title{Extracts the drug carriers element and return data as data frame.}
 \usage{
-parse_drug_carriers(save_table = FALSE, save_csv = FALSE,
-  csv_path = ".", override_csv = FALSE)
+carriers(save_table = FALSE, save_csv = FALSE, csv_path = ".",
+  override_csv = FALSE)
 }
 \arguments{
 \item{save_table}{boolean, save table in database if true.}
@@ -22,7 +22,7 @@ in the new parse operation}
 drug carriers node attributes date frame
 }
 \description{
-\code{parse_drug_carriers} returns data frame of drug carriers elements.
+\code{carriers} returns data frame of drug carriers elements.
 }
 \details{
 This functions extracts the carriers element of drug node in \strong{DrugBank}
@@ -37,30 +37,30 @@ no need to call it again before calling this function.
 \examples{
 \dontrun{
 # return only the parsed dataframe
-parse_drug_carriers()
+carriers()
 
 # save in database and return parsed dataframe
-parse_drug_carriers(save_table = TRUE)
+carriers(save_table = TRUE)
 
 # save parsed dataframe as csv if it does not exist in current
 # location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_drug_carriers(save_csv = TRUE)
+carriers(save_csv = TRUE)
 
 # save in database, save parsed dataframe as csv if it does not exist
 # in current location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_drug_carriers(ssave_table = TRUE, save_csv = TRUE)
+carriers(ssave_table = TRUE, save_csv = TRUE)
 
 # save parsed dataframe as csv if it does not exist in given
 # location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-parse_drug_carriers(save_csv = TRUE, csv_path = TRUE)
+carriers(save_csv = TRUE, csv_path = TRUE)
 
 # save parsed dataframe as csv if it does not exist in current
 # location and return parsed dataframe.
 # If the csv exist override it and return it.
-parse_drug_carriers(save_csv = TRUE, csv_path = TRUE, override = TRUE)
+carriers(save_csv = TRUE, csv_path = TRUE, override = TRUE)
 }
 }
 \seealso{
