+++ R/drug_sequence_node_parser.R
@@ -25,7 +25,8 @@ get_sequences_df <- function(rec) {
 #'
 #' \code{parse_drug_sequences} returns data frame of drug sequences elements.
 #'
-#' This functions extracts the sequences element of drug node in \strong{DrugBank}
+#' This functions extracts the sequences element of drug node in
+#' \strong{DrugBank}
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. It takes one single optional argument to
 #' save the returned dataframe in the database.
