+++ R/drug_carrier_node_parser.R
@@ -322,9 +322,10 @@ carriers_textbooks <-
 #'
 #' \code{parse_drug_groups} returns data frame of drug carriers links elements.
 #'
-#' This functions extracts the carriers links element of drug node in \strong{DrugBank}
-#' xml database with the option to save it in a predefined database via
-#' \code{\link{open_db}} method. It takes one single optional argument to
+#' This functions extracts the carriers links element of drug node in
+#' \strong{DrugBank} xml database with the option to save it in a predefined
+#' database via \code{\link{open_db}} method.
+#' It takes one single optional argument to
 #' save the returned dataframe in the database.
 #' It must be called after \code{\link{read_drugbank_xml_db}} function like
 #' any other parser function.
@@ -895,7 +896,8 @@ carriers_polypeptidepeptides_go <-
 #'
 #' \code{carriers} returns data frame of drug carriers elements.
 #'
-#' This functions extracts the carriers element of drug node in \strong{DrugBank}
+#' This functions extracts the carriers element of drug node in
+#'  \strong{DrugBank}
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. It takes one single optional argument to
 #' save the returned dataframe in the database.
