+++ R/drug_manufacturer_node_parser.R
@@ -20,7 +20,8 @@ get_manufactures_df <- function(rec) {
 #' \code{parse_drug_manufacturers} returns data frame of drug manufacturers
 #' elements.
 #'
-#' This functions extracts the manufacturers element of drug node in \strong{DrugBank}
+#' This functions extracts the manufacturers element of drug node in
+#'  \strong{DrugBank}
 #' xml database with the option to save it in a predefined database via
 #' \code{\link{open_db}} method. It takes one single optional argument to
 #' save the returned dataframe in the database.

[…]
