+++ R/drug_enzyme_node_parser.R
@@ -89,7 +89,7 @@ get_enzy_poly_go_df <- function(rec) {
 
 #' Extracts the drug enzymes actions element and return data as data frame.
 #'
-#' \code{parse_drug_enzymes_actions} returns data frame of drug enzymes
+#' \code{enzymes_actions} returns data frame of drug enzymes
 #' actions elements.
 #'
 #' This functions extracts the enzymes actions element of drug node in
@@ -109,37 +109,37 @@ get_enzy_poly_go_df <- function(rec) {
 #' @param override_csv override existing csv, if any, in case it is true in the
 #'  new parse operation
 #' @return drug enzymes actions node attributes date frame
-#'
+#' @family enzymes
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_drug_enzymes_actions()
+#' enzymes_actions()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug_enzymes_actions(save_table = TRUE)
+#' enzymes_actions(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_enzymes_actions(save_csv = TRUE)
+#' enzymes_actions(save_csv = TRUE)
 #'
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_enzymes_actions(ssave_table = TRUE, save_csv = TRUE)
+#' enzymes_actions(ssave_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given
 #' # location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_enzymes_actions(save_csv = TRUE, csv_path = TRUE)
+#' enzymes_actions(save_csv = TRUE, csv_path = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist override it and return it.
-#' parse_drug_enzymes_actions(save_csv = TRUE, csv_path = TRUE, override = TRUE)
+#' enzymes_actions(save_csv = TRUE, csv_path = TRUE, override = TRUE)
 #' }
 #' @export
-parse_drug_enzymes_actions <-
+enzymes_actions <-
   function(save_table = FALSE,
            save_csv = FALSE,
            csv_path = ".",
@@ -173,7 +173,7 @@ parse_drug_enzymes_actions <-
 
 #' Extracts the drug enzymes articles element and return data as data frame.
 #'
-#' \code{parse_drug_enzymes_articles} returns data frame of drug enzymes
+#' \code{enzymes_articles} returns data frame of drug enzymes
 #' articles elements.
 #'
 #' This functions extracts the enzymes articles element of drug node in drugbank
@@ -192,40 +192,40 @@ parse_drug_enzymes_actions <-
 #' @param override_csv override existing csv, if any, in case it is true
 #' in the new parse operation
 #' @return drug enzymes articles node attributes date frame
-#'
+#' @family enzymes
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_drug_enzymes_articles()
+#' enzymes_articles()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug_enzymes_articles(save_table = TRUE)
+#' enzymes_articles(save_table = TRUE)
 #'
[…]
