### 2019-12-23 3c3cf0c
#74, change target functions
