+++ R/drug_targets_node_parser.R
@@ -62,7 +62,7 @@ get_targ_poly_go_df <- function(rec) {
 #' Extracts the drug targ polypeptides external identifiers
 #'  element and return data as data frame.
 #'
-#' \code{parse_targ_poly_ext_identity}
+#' \code{targets_polypeptide_ext_ident}
 #'  returns data frame of drug targ polypeptides external identifiers
 #'  elements.
 #'
@@ -84,25 +84,25 @@ get_targ_poly_go_df <- function(rec) {
 #'  new parse operation
 #' @return drug targ polypeptides external identifiers node attributes
 #' date frame
-#'
+#' @family targets
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_targ_poly_ext_identity()
+#' targets_polypeptide_ext_ident()
 #'
 #' # save in database and return parsed dataframe
-#' parse_targ_poly_ext_identity(save_table = TRUE)
+#' targets_polypeptide_ext_ident(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current location
 #' # and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_targ_poly_ext_identity(save_csv = TRUE)
+#' targets_polypeptide_ext_ident(save_csv = TRUE)
 #'
 #' # save in database, save parsed dataframe as csv if it does not exist in
 #' current
 #' # location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_targ_poly_ext_identity(
+#' targets_polypeptide_ext_ident(
 #'   ssave_table = TRUE,
 #'   save_csv = TRUE
 #' )
@@ -110,7 +110,7 @@ get_targ_poly_go_df <- function(rec) {
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_targ_poly_ext_identity(
+#' targets_polypeptide_ext_ident(
 #'   save_csv = TRUE,
 #'   csv_path = TRUE
 #' )
@@ -118,12 +118,12 @@ get_targ_poly_go_df <- function(rec) {
 #' # save parsed dataframe as csv if it does not exist in current location
 #' # and return parsed dataframe.
 #' # If the csv exist override it and return it.
-#' parse_targ_poly_ext_identity(
+#' targets_polypeptide_ext_ident(
 #'   save_csv = TRUE, csv_path = TRUE, override = TRUE
 #' )
 #' }
 #' @export
-parse_targ_poly_ext_identity <-
+targets_polypeptide_ext_ident <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".",
            override_csv = FALSE) {
     path <-
@@ -162,7 +162,7 @@ parse_targ_poly_ext_identity <-
 #' Extracts the drug targ polypeptides syn element
 #' and return data as data frame.
 #'
-#' \code{parse_targ_poly_syn} returns data
+#' \code{targets_polypeptide} returns data
 #'  frame of drug targ polypeptides syn elements.
 #'
 #' This functions extracts the targ polypeptides syn element of
@@ -182,40 +182,40 @@ parse_targ_poly_ext_identity <-
 #' @param override_csv override existing csv, if any, in case it is true in the
 #'  new parse operation
 #' @return drug targ polypeptides syn node attributes date frame
-#'
+#' @family targets
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_targ_poly_syn()
+#' targets_polypeptide()
 #'
 #' # save in database and return parsed dataframe
-#' parse_targ_poly_syn(save_table = TRUE)
+#' targets_polypeptide(save_table = TRUE)
 #'
[…]
