### 2019-12-24 d473590
#74, fix wrongly renamed method
