+++ R/drug_transporter_node_parser.R
@@ -68,7 +68,7 @@ get_trans_poly_go_df <- function(rec) {
 
 #' Extracts the drug transporters actions element and return data as data frame.
 #'
-#' \code{parse_drug_transporters actions} returns data frame of drug
+#' \code{transporters actions} returns data frame of drug
 #'  transporters
 #' actions elements.
 #'
@@ -89,40 +89,40 @@ get_trans_poly_go_df <- function(rec) {
 #' @param override_csv override existing csv, if any, in case it is true in
 #'  the new parse operation
 #' @return drug transporters actions node attributes date frame
-#'
+#' @family transporters
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_drug_trans_actions()
+#' transporters_actions()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug_trans_actions(save_table = TRUE)
+#' transporters_actions(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current location
 #' # and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_trans_actions(save_csv = TRUE)
+#' transporters_actions(save_csv = TRUE)
 #'
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_trans_actions(ssave_table = TRUE, save_csv = TRUE)
+#' transporters_actions(ssave_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_trans_actions(save_csv = TRUE, csv_path = TRUE)
+#' transporters_actions(save_csv = TRUE, csv_path = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current location
 #' #  and return parsed dataframe.
 #' # If the csv exist override it and return it.
-#' parse_drug_trans_actions(
+#' transporters_actions(
 #'   save_csv = TRUE, csv_path = TRUE,
 #'   override = TRUE
 #' )
 #' }
 #' @export
-parse_drug_trans_actions <-
+transporters_actions <-
   function(save_table = FALSE,
            save_csv = FALSE,
            csv_path = ".",
@@ -159,7 +159,7 @@ parse_drug_trans_actions <-
 #' Extracts the drug transporters articles element and return data as
 #' data frame.
 #'
-#' \code{parse_drug_transporters articles} returns data frame of drug
+#' \code{transporters articles} returns data frame of drug
 #' transporters articles elements.
 #'
 #' This functions extracts the transporters articles element of drug node in
@@ -179,40 +179,40 @@ parse_drug_trans_actions <-
 #' @param override_csv override existing csv, if any, in case it is true in the
 #'  new parse operation
 #' @return drug transporters articles node attributes date frame
-#'
+#' @family transporters
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_drug_trans_articles()
+#' transporters_articles()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug_trans_articles(save_table = TRUE)
+#' transporters_articles(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current location
 #' # and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
[…]
