+++ tests/testthat/test_drug_transporter_node_parser_biotech.R
@@ -22,10 +22,10 @@ test_that(
   desc = "Read darug transporters actions attributes",
   code = {
     expect_equal(
-      nrow(parse_drug_trans_actions()),
+      nrow(transporters_actions()),
       0
     )
-    expect_error(parse_drug_trans_actions(TRUE))
+    expect_error(transporters_actions(TRUE))
   }
 )
 
@@ -33,10 +33,10 @@ test_that(
   desc = "Read darug transporters_articles attributes",
   code = {
     expect_equal(
-      nrow(parse_drug_trans_articles()),
+      nrow(transporters_articles()),
       0
     )
-    expect_error(parse_drug_trans_articlese(TRUE))
+    expect_error(transporters_articlese(TRUE))
   }
 )
 
@@ -44,10 +44,10 @@ test_that(
   desc = "Read darug transporters textbooks attributes",
   code = {
     expect_equal(
-      nrow(parse_drug_trans_textbooks()),
+      nrow(transporters_textbooks()),
       0
     )
-    expect_error(parse_drug_trans_textbooks(TRUE))
+    expect_error(transporters_textbooks(TRUE))
   }
 )
 
@@ -55,10 +55,10 @@ test_that(
   desc = "Read darug transporters links attributes",
   code = {
     expect_equal(
-      nrow(parse_drug_trans_links()),
+      nrow(transporters_links()),
       0
     )
-    expect_error(parse_drug_trans_links(TRUE))
+    expect_error(transporters_links(TRUE))
   }
 )
 
@@ -66,10 +66,10 @@ test_that(
   desc = "Read darug transporters polypeptides attributes",
   code = {
     expect_equal(
-      nrow(parse_drug_trans_polys()),
+      nrow(transporters_polypeptide()),
       0
     )
-    expect_error(parse_drug_trans_polys(TRUE))
+    expect_error(transporters_polypeptide(TRUE))
   }
 )
 
@@ -77,11 +77,11 @@ test_that(
   desc = "Read darug transporters polypeptides external identifiers attributes",
   code = {
     expect_equal(
-      nrow(parse_trans_poly_ex_identity()),
+      nrow(transporters_polypep_ex_ident()),
       0
     )
     expect_error(
-      parse_trans_poly_ex_identity(TRUE)
+      transporters_polypep_ex_ident(TRUE)
     )
   }
 )
@@ -90,10 +90,10 @@ test_that(
   desc = "Read darug transporters polypeptides syn attributes",
   code = {
     expect_equal(
-      nrow(parse_trans_poly_syn()),
+      nrow(transporters_polypeptide_syn()),
       0
     )
-    expect_error(parse_trans_poly_syn(TRUE))
+    expect_error(transporters_polypeptide_syn(TRUE))
   }
 )
 
@@ -101,10 +101,10 @@ test_that(
   desc = "Read darug transporters polypeptides pfams attributes",
   code = {
     expect_equal(
-      nrow(parse_trans_poly_pfams()),
+      nrow(transporters_polypeptide_pfams()),
       0
     )
-    expect_error(parse_trans_poly_pfams(TRUE))
+    expect_error(transporters_polypeptide_pfams(TRUE))
   }
 )
 
@@ -112,10 +112,10 @@ test_that(
   desc = "Read darug transporters polypeptides go classifiers attributes",
   code = {
     expect_equal(
-      nrow(parse_trans_poly_go()),
+      nrow(transporters_polypeptide_go()),
       0
     )
-    expect_error(parse_trans_poly_go(TRUE))
+    expect_error(transporters_polypeptide_go(TRUE))
   }
 )
 
@@ -123,9 +123,9 @@ test_that(
   desc = "Read darug transporters attributes",
[…]
