+++ man/targets_links.Rd
@@ -71,6 +71,7 @@ Other targets: \code{\link{targets_actions}},
   \code{\link{targets_polypeptide_ext_ident}},
   \code{\link{targets_polypeptide_go}},
   \code{\link{targets_polypeptide_pfams}},
+  \code{\link{targets_polypeptide_syn}},
   \code{\link{targets_polypeptide}},
   \code{\link{targets_textbooks}}, \code{\link{targets}}
 }

[…]
