+++ man/drug_element.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/drug_all_nodes_parser.R
-\name{parse_drug_element}
-\alias{parse_drug_element}
+\name{drug_element}
+\alias{drug_element}
 \title{extracts the given drug elements and return data as list of dataframes.}
 \usage{
-parse_drug_element(elements_options = c("all"), save_table = FALSE,
+drug_element(elements_options = c("all"), save_table = FALSE,
   save_csv = FALSE, csv_path = ".", override_csv = FALSE)
 }
 \arguments{
@@ -25,7 +25,7 @@ new parse operation}
 list of selected drug elements dataframes
 }
 \description{
-\code{parse_drug_element} returns list of dataframes of drugs selected
+\code{drug_element} returns list of dataframes of drugs selected
 elements.
 }
 \details{
@@ -38,38 +38,43 @@ any other parser function.
 if \code{\link{read_drugbank_xml_db}} is called before for any reason, so
 no need to call it again before calling this function.
 
-parse_drug_element_options can be called to know the valid options for
+drug_element_options can be called to know the valid options for
 this method
 }
 \examples{
 \dontrun{
 # return only the parsed dataframe
-parse_drug_element()
+drug_element()
 
 # save in database and return parsed dataframe
-parse_drug_element(save_table = trUe)
+drug_element(save_table = trUe)
 
 # save parsed dataframe as csv if it does not exist in current location and
 # return parsed dataframe.
 # if the csv exist before read it and return its data.
-parse_drug_element(save_csv = trUe)
+drug_element(save_csv = trUe)
 
 # save in database, save parsed dataframe as csv if it does not
 # exist in current location and return parsed dataframe.
 # if the csv exist before read it and return its data.
-parse_drug_element(ssave_table = trUe, save_csv = trUe)
+drug_element(ssave_table = trUe, save_csv = trUe)
 
 # save parsed dataframe as csv if it does not exist in given location and
 # return parsed dataframe.
 # if the csv exist before read it and return its data.
-parse_drug_element(save_csv = trUe, csv_path = trUe)
+drug_element(save_csv = trUe, csv_path = trUe)
 
 # save parsed dataframe as csv if it does not exist in current
 # location and return parsed dataframe.
 # if the csv exist override it and return it.
-parse_drug_element(save_csv = trUe, csv_path = trUe, override = trUe)
-parse_drug_element(c("drug_ahfs_codes", "drug_carriers"), save_table = trUe)
-parse_drug_element(save_table = FALSE)
-parse_drug_element(c("drug_ahfs_codes", "drug_carriers"))
+drug_element(save_csv = trUe, csv_path = trUe, override = trUe)
+drug_element(c("drug_ahfs_codes", "drug_carriers"), save_table = trUe)
+drug_element(save_table = FALSE)
+drug_element(c("drug_ahfs_codes", "drug_carriers"))
 }
 }
+\seealso{
+Other common: \code{\link{drug_all}},
+  \code{\link{drug_element_options}}
+}
+\concept{common}
