+++ man/parse_drug_element_options.Rd
@@ -1,19 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/drug_all_nodes_parser.R
-\name{parse_drug_element_options}
-\alias{parse_drug_element_options}
-\title{returns \code{parse_drug_element} valid options.}
-\usage{
-parse_drug_element_options()
-}
-\value{
-list of \code{parse_drug_element} valid options
-}
-\description{
-returns \code{parse_drug_element} valid options.
-}
-\examples{
-\dontrun{
-parse_drug_element_options()
-}
-}
