+++ R/drug_parsers.R
@@ -1,6 +1,6 @@
 #' Extracts the main drug elements and return data as tibble.
 #'
-#' \code{parse_drug} returns tibble of drugs main elements.
+#' \code{drug} returns tibble of drugs main elements.
 #'
 #' This functions extracts the main element of drug node in drugbank
 #' xml database with the option to save it in a predefined database via
@@ -18,37 +18,37 @@
 #' @param override_csv override existing csv, if any, in case it is true in the
 #'  new parse operation
 #' @return drug main node attributes tibble
-#'
+#' @family drugs
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_drug()
+#' drug()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug(save_table = TRUE)
+#' drug(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug(save_csv = TRUE)
+#' drug(save_csv = TRUE)
 #'
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug(ssave_table = TRUE, save_csv = TRUE)
+#' drug(ssave_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' #  and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug(save_csv = TRUE, csv_path = TRUE)
+#' drug(save_csv = TRUE, csv_path = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist override it and return it.
-#' parse_drug(save_csv = TRUE, csv_path = TRUE, override = TRUE)
+#' drug(save_csv = TRUE, csv_path = TRUE, override = TRUE)
 #' }
 #' @export
-parse_drug <-
+drug <-
   function(save_table = FALSE,
            save_csv = FALSE,
            csv_path = ".",
@@ -150,7 +150,7 @@ parse_drug <-
 
 #' Extracts the drug groups element and return data as tibble.
 #'
-#' \code{parse_drug_groups} returns tibble of drug groups elements.
+#' \code{drug_groups} returns tibble of drug groups elements.
 #'
 #' This functions extracts the groups element of drug node in drugbank
 #' xml database with the option to save it in a predefined database via
@@ -168,37 +168,37 @@ parse_drug <-
 #' @param override_csv override existing csv, if any, in case it is true in the
 #'  new parse operation
 #' @return drug groups node attributes tibble
-#'
+#' @family drugs
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_drug_groups()
+#' drug_groups()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug_groups(save_table = TRUE)
+#' drug_groups(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_groups(save_csv = TRUE)
+#' drug_groups(save_csv = TRUE)
 #'
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_groups(ssave_table = TRUE, save_csv = TRUE)
[…]
