+++ R/drug_price_node_parser.R
@@ -18,7 +18,7 @@ get_prices_df <- function(rec) {
 
 #' Extracts the drug prices element and return data as data frame.
 #'
-#' \code{parse_drug_prices} returns data frame of drug prices elements.
+#' \code{drug_prices} returns data frame of drug prices elements.
 #'
 #' This functions extracts the prices element of drug node in \strong{DrugBank}
 #' xml database with the option to save it in a predefined database via
@@ -40,33 +40,33 @@ get_prices_df <- function(rec) {
 #' @examples
 #' \dontrun{
 #' # return only the parsed dataframe
-#' parse_drug_prices()
+#' drug_prices()
 #'
 #' # save in database and return parsed dataframe
-#' parse_drug_prices(save_table = TRUE)
+#' drug_prices(save_table = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_prices(save_csv = TRUE)
+#' drug_prices(save_csv = TRUE)
 #'
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_prices(ssave_table = TRUE, save_csv = TRUE)
+#' drug_prices(ssave_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' parse_drug_prices(save_csv = TRUE, csv_path = TRUE)
+#' drug_prices(save_csv = TRUE, csv_path = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in current
 #' # location and return parsed dataframe.
 #' # If the csv exist override it and return it.
-#' parse_drug_prices(save_csv = TRUE, csv_path = TRUE, override = TRUE)
+#' drug_prices(save_csv = TRUE, csv_path = TRUE, override = TRUE)
 #' }
 #' @export
-parse_drug_prices <- function(save_table = FALSE, save_csv = FALSE,
+drug_prices <- function(save_table = FALSE, save_csv = FALSE,
                               csv_path = ".", override_csv = FALSE) {
   path <-
     get_dataset_full_path("drug_prices", csv_path)
