### 2019-12-24 79d8a1d
fix #74, change last method name
