### 2019-12-26 a32ac1e
#14, fix eploration section
