+++ README.md
@@ -50,20 +50,20 @@ This is a basic example which shows you how to solve a common problem:
 
 ``` r
 ## parse data from XML and save it to memory
-dbparser::get_xml_db_rows(
+dbparser::read_drugbank_xml_db(
   "D:/DS Projects/Drug_Interaction/drugbank.xml"
               #system.file("extdata", "drugbank_record.xml", package = "dbparser")
             )
 #> [1] TRUE
 
 ## load drugs data
-drugs <- dbparser::parse_drug()
+drugs <- dbparser::drug()
 
 ## load drug groups data
-drug_groups <- dbparser::parse_drug_groups()
+drug_groups <- dbparser::drug_groups()
 
 ## load drug targets actions data
-drug_targets_actions <- dbparser::parse_drug_targets_actions()
+drug_targets_actions <- dbparser::targets_actions()
 ```
 
 ## Saving into a database
@@ -94,75 +94,6 @@ library(dbparser)
  close_db()
 ```
 
-## Exploring the data
-
-Following is an example involving a quick look at a few aspects of the
-parsed data. First we look at the proportions of `biotech` and
-`small-molecule` drugs in the
-data.
-
-``` r
-## view proportions of the different drug types (biotech vs. small molecule)
-drugs %>% 
-    select(type) %>% 
-    ggplot(aes(x = type, fill = type)) + 
-    geom_bar() + 
-    guides(fill=FALSE)     ## removes legend for the bar colors
-```
-
-<img src="man/figures/README-unnamed-chunk-2-1.png" width="100%" />
-
-Below, we view the different `drug_groups` in the data and how prevalent
-they are.
-
-``` r
-## view proportions of the different drug types for each drug group
-drugs %>% 
-    full_join(drug_groups, by = c('primary_key' = 'drugbank_id')) %>% 
-    select(type, group) %>% 
-    ggplot(aes(x = group, fill = type)) + 
-    geom_bar() + 
-    theme(legend.position= 'bottom') + 
-    labs(x = 'Drug Group', 
-         y = 'Quantity', 
-         title="Drug Type Distribution per Drug Group", 
-         caption="created by ggplot") + 
-    coord_flip()
-```
-
-<img src="man/figures/README-unnamed-chunk-3-1.png" width="100%" />
-
-Finally, we look at the `drug_targets_actions` to observe their
-proportions as well.
-
-``` r
-## get counts of the different target actions in the data
-targetActionCounts <- 
-    drug_targets_actions %>% 
-    group_by(action) %>% 
-    summarise(count = n()) %>% 
-    arrange(desc(count))
-
-## get bar chart of the 10 most occurring target actions in the data
-p <- 
-    ggplot(targetActionCounts[1:10,], 
-           aes(x = reorder(action,count), y = count, fill = letters[1:10])) + 
-    geom_bar(stat = 'identity') +
-    labs(fill = 'action', 
-         x = 'Target Action', 
-         y = 'Quantity', 
-         title = 'Target Actions Distribution', 
-         subtitle = 'Distribution of Target Actions in the Data',
-         caption = 'created by ggplot') + 
-    guides(fill = FALSE) +    ## removes legend for the bar colors
-    coord_flip()              ## switches the X and Y axes
-
-## display plot
-p
-```
-
-<img src="man/figures/README-unnamed-chunk-4-1.png" width="100%" />
-
 ## Code of Conduct
 
 Please note that the ‘dbparser’ project is released with a [Contributor
@@ -186,7 +117,7 @@ citation("dbparser")
 #> 
 #> To cite dbparser in publications use:
 #> 
[…]
