### 2019-12-26 79ee9b2
#14, fix issues errors
