### 2019-12-31 07efa53
#14, update data exploraion section
