### 2019-12-31 96c2732
#14, update exploration tutorial
