+++ vignettes/dbparser.Rmd
@@ -21,12 +21,13 @@ knitr::opts_chunk$set(
 The main purpose of the `dbparser` package is to parse the 
 [DrugBank](http://drugbank.ca/) database which is downloadable in XML format 
 from [this link](https://www.drugbank.ca/releases/latest). The parsed data can 
-then be explored and analyzed as desired by the user. The `dbparser` package 
-further provides the facility of saving the parsed data into a given database.
+then be explored and analyzed as desired by the user. 
+In this tutorial, we will see how to use `dbparser` along with `dplyr` and 
+`ggplot2` along with other libraries to do simple drug analyasis.
 
 
 
-## Getting Started -- Loading the Data
+## Loading and Parsing the Data
 
 Following is sample code that parses the DrugBank database, then loads the 
 drugs info, drug groups info and drug targets actions info.
@@ -67,7 +68,6 @@ drugs %>%
     guides(fill = FALSE)     ## removes legend for the bar colors
 ```
 
-image: ![](fig1.png)
 
 Below, we view the different `drug_groups` in the data and how prevalent they 
 are.
@@ -90,7 +90,7 @@ drugs %>%
 Finally, we look at the `drug_targets_actions` to observe their proportions as 
 well.
 
-image: ![](fig2.png)
+![](fig2.png)
 
 ```{r eval=FALSE}
 ## get counts of the different target actions in the data
@@ -118,4 +118,4 @@ p <-
 p
 ```
 
-image: ![](fig3.png)
+![](fig3.png)
