+++ vignettes/dbparser.Rmd
@@ -18,14 +18,13 @@ knitr::opts_chunk$set(
 )
 ```
 
+## Introduction
 The main purpose of the `dbparser` package is to parse the 
 [DrugBank](http://drugbank.ca/) database which is downloadable in XML format 
 from [this link](https://www.drugbank.ca/releases/latest). The parsed data can 
 then be explored and analyzed as desired by the user. 
 In this tutorial, we will see how to use `dbparser` along with `dplyr` and 
 `ggplot2` along with other libraries to do simple drug analyasis.
-
-
 
 ## Loading and Parsing the Data
 
