### 2019-12-31 94e9152
#14, fix function documentation word errors
