+++ R/drug_parsers.R
@@ -35,7 +35,7 @@
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug(ssave_table = TRUE, save_csv = TRUE)
+#' drug(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' #  and return parsed dataframe.
@@ -185,7 +185,7 @@ drug <-
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_groups(ssave_table = TRUE, save_csv = TRUE)
+#' drug_groups(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given
 #' # location and return parsed dataframe.
@@ -264,7 +264,7 @@ drug_groups <-
 #' # save in database, save parsed dataframe as csv if it
 #' # does not exist in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_articles(ssave_table = TRUE, save_csv = TRUE)
+#' drug_articles(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given
 #' # location and return parsed dataframe.
@@ -344,7 +344,7 @@ drug_articles <-
 #' # save in database, save parsed dataframe as csv if it does not
 #' # exist in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_books(ssave_table = TRUE, save_csv = TRUE)
+#' drug_books(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given l
 #' # ocation and return parsed dataframe.
@@ -422,7 +422,7 @@ drug_books <-
 #' # save in database, save parsed dataframe as csv if it does not
 #' # exist in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_links(ssave_table = TRUE, save_csv = TRUE)
+#' drug_links(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given
 #' # location and return parsed dataframe.
@@ -501,7 +501,7 @@ drug_links <-
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_syn(ssave_table = TRUE, save_csv = TRUE)
+#' drug_syn(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.
@@ -577,7 +577,7 @@ drug_syn <-
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_products(ssave_table = TRUE, save_csv = TRUE)
+#' drug_products(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.
@@ -654,7 +654,7 @@ drug_products <-
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_calc_prop(ssave_table = TRUE, save_csv = TRUE)
[…]
