+++ R/drug_atc_codes_node_parser.R
@@ -63,7 +63,7 @@ get_atc_codes_df <- function(rec) {
 #' current
 #' #  location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_atc_codes(ssave_table = TRUE, save_csv = TRUE)
+#' drug_atc_codes(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location and
 #' # return parsed dataframe.
