+++ R/drug_sequence_node_parser.R
@@ -59,7 +59,7 @@ get_sequences_df <- function(rec) {
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_sequences(ssave_table = TRUE, save_csv = TRUE)
+#' drug_sequences(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.
