+++ R/drug_manufacturer_node_parser.R
@@ -54,7 +54,7 @@ get_manufactures_df <- function(rec) {
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_manufacturers(ssave_table = TRUE, save_csv = TRUE)
+#' drug_manufacturers(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.
