+++ R/drug_classfication_node_parser.R
@@ -49,7 +49,7 @@ drug_classifications_df <- function(rec) {
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_classification(ssave_table = TRUE, save_csv = TRUE)
+#' drug_classification(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' #  and return parsed dataframe.
