+++ R/drug_price_node_parser.R
@@ -53,7 +53,7 @@ get_prices_df <- function(rec) {
 #' # save in database, save parsed dataframe as csv if it does not exist
 #' # in current location and return parsed dataframe.
 #' # If the csv exist before read it and return its data.
-#' drug_prices(ssave_table = TRUE, save_csv = TRUE)
+#' drug_prices(save_table = TRUE, save_csv = TRUE)
 #'
 #' # save parsed dataframe as csv if it does not exist in given location
 #' # and return parsed dataframe.

[…]
