### 2020-01-01 d3140a5
fix #76, add database connection checker
