+++ R/drug_all_nodes_parser.R
@@ -51,6 +51,7 @@ drug_all <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     drugs <- drug(save_table, save_csv, csv_path, override_csv)
     message("parsed drugs main attributes, 1/75")
     groups_drug <-
@@ -123,7 +124,7 @@ drug_all <-
       drug_snp_effects(save_table, save_csv, csv_path, override_csv)
     message("parsed snp_effects_drug, 21/75")
     snp_adverse_reactions <-
-      drug_snp_adverse_reactions (
+      drug_snp_adverse_reactions(
         save_table, save_csv, csv_path,
         override_csv
       )
@@ -528,6 +529,7 @@ drug_element <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     if (!all(elements_options %in% drug_element_options())) {
       stop("invalid options\nplease use drug_element_options() to
            know valid options")
