+++ R/drug_targets_node_parser.R
@@ -126,6 +126,7 @@ get_targ_poly_go_df <- function(rec) {
 targets_polypeptide_ext_ident <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <-
       get_dataset_full_path(
         "drug_targ_poly_ex_identity",
@@ -218,6 +219,7 @@ targets_polypeptide_ext_ident <-
 targets_polypeptide_syn <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <-
       get_dataset_full_path("drug_targ_poly_syn", csv_path)
     if (!override_csv & file.exists(path)) {
@@ -305,6 +307,7 @@ targets_polypeptide_syn <-
 targets_polypeptide_pfams <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <-
       get_dataset_full_path("drug_targ_poly_pfams", csv_path)
     if (!override_csv & file.exists(path)) {
@@ -398,6 +401,7 @@ targets_polypeptide_pfams <-
 targets_polypeptide_go <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <-
       get_dataset_full_path(
         "drug_targ_polys_go",
@@ -481,6 +485,7 @@ targets_polypeptide_go <-
 #' @export
 targets_actions <- function(save_table = FALSE, save_csv = FALSE,
                                     csv_path = ".", override_csv = FALSE) {
+  check_database_connection(save_table)
   path <-
     get_dataset_full_path("drug_targ_actions", csv_path)
   if (!override_csv & file.exists(path)) {
@@ -565,6 +570,7 @@ targets_actions <- function(save_table = FALSE, save_csv = FALSE,
 #' @export
 targets_articles <- function(save_table = FALSE, save_csv = FALSE,
                                      csv_path = ".", override_csv = FALSE) {
+  check_database_connection(save_table)
   path <-
     get_dataset_full_path("drug_targ_articles", csv_path)
   if (!override_csv & file.exists(path)) {
@@ -644,6 +650,7 @@ targets_articles <- function(save_table = FALSE, save_csv = FALSE,
 #' @export
 targets_textbooks <- function(save_table = FALSE, save_csv = FALSE,
                                       csv_path = ".", override_csv = FALSE) {
+  check_database_connection(save_table)
   path <-
     get_dataset_full_path("drug_targ_textbooks", csv_path)
   if (!override_csv & file.exists(path)) {
@@ -721,6 +728,7 @@ targets_textbooks <- function(save_table = FALSE, save_csv = FALSE,
 #' @export
 targets_links <- function(save_table = FALSE, save_csv = FALSE,
                                   csv_path = ".", override_csv = FALSE) {
+  check_database_connection(save_table)
   path <-
     get_dataset_full_path("drug_targ_links", csv_path)
   if (!override_csv & file.exists(path)) {
@@ -814,6 +822,7 @@ targets_links <- function(save_table = FALSE, save_csv = FALSE,
 targets_polypeptide <- function(save_table = FALSE,
                                   save_csv = FALSE, csv_path = ".",
                                   override_csv = FALSE) {
+  check_database_connection(save_table)
   path <-
     get_dataset_full_path("drug_targ_polys", csv_path)
   if (!override_csv & file.exists(path)) {
[…]
