+++ R/drug_parsers.R
@@ -53,6 +53,7 @@ drug <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <- get_dataset_full_path("drugs", csv_path)
     if (!override_csv & file.exists(path)) {
       drugs <- readr::read_csv(path)
@@ -203,6 +204,7 @@ drug_groups <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <- get_dataset_full_path("drug_groups", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_groups <- readr::read_csv(path)
@@ -282,6 +284,7 @@ drug_articles <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <- get_dataset_full_path("drug_articles", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_articles <- readr::read_csv(path)
@@ -362,6 +365,7 @@ drug_books <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <- get_dataset_full_path("drug_books", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_books <- readr::read_csv(path)
@@ -440,6 +444,7 @@ drug_links <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <- get_dataset_full_path("drug_links", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_links <- readr::read_csv(path)
@@ -519,6 +524,7 @@ drug_syn <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <- get_dataset_full_path("drug_syn", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_syn <- readr::read_csv(path)
@@ -595,6 +601,7 @@ drug_products <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <- get_dataset_full_path("drug_products", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_products <- readr::read_csv(path)
@@ -674,6 +681,7 @@ drug_calc_prop <- function(save_table = FALSE,
                                  save_csv = FALSE,
                                  csv_path = ".",
                                  override_csv = FALSE) {
+  check_database_connection(save_table)
   path <-
     get_dataset_full_path("drug_calculated_properties", csv_path)
   if (!override_csv & file.exists(path)) {
@@ -759,6 +767,7 @@ drug_intern_brand <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <-
       get_dataset_full_path("drug_international_brands", csv_path)
     if (!override_csv & file.exists(path)) {
@@ -839,6 +848,7 @@ drug_salts <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
     path <-
       get_dataset_full_path("drug_salts", csv_path)
     if (!override_csv & file.exists(path)) {
@@ -916,6 +926,7 @@ drug_mixtures <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
+    check_database_connection(save_table)
[…]
