+++ R/drug_classfication_node_parser.R
@@ -64,6 +64,7 @@ drug_classifications_df <- function(rec) {
 #' @export
 drug_classification <- function(save_table = FALSE, save_csv = FALSE,
                                       csv_path = ".", override_csv = FALSE) {
+  check_database_connection(save_table)
   path <-
     get_dataset_full_path("drug_classifications", csv_path)
   if (!override_csv & file.exists(path)) {
