+++ R/drug_sequence_node_parser.R
@@ -74,6 +74,7 @@ get_sequences_df <- function(rec) {
 #' @export
 drug_sequences <- function(save_table = FALSE, save_csv = FALSE,
                                  csv_path = ".", override_csv = FALSE) {
+  check_database_connection(save_table)
   path <-
     get_dataset_full_path("drug_sequences", csv_path)
   if (!override_csv & file.exists(path)) {

[…]
