+++ DESCRIPTION
@@ -4,7 +4,9 @@ Version: 1.0.5.9000
 Authors@R:
     c(
     person("Mohammed", "Ali", email = "moh_fcis@yahoo.com", role = c("aut", "cre")),
-    person("Ali", "Ezzat", email = "ali_ezzat85@yahoo.com", role = "aut"))
+    person("Ali", "Ezzat", email = "ali_ezzat85@yahoo.com", role = "aut"),
+    person("Hao", "Zhu", email = "haozhu233@gmail.com", role = "rev"),
+    person("Emma", "Mendelsohn", email = "", role = "rev"))
 Description: This tool is for parsing the 'DrugBank' XML database <http://drugbank.ca/>. The parsed 
     data are then returned in a proper 'R' dataframe with the ability to save 
     them in a given database.
