+++ man/drug_atc_codes.Rd
@@ -52,7 +52,7 @@ drug_atc_codes(save_csv = TRUE)
 current
 #  location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-drug_atc_codes(ssave_table = TRUE, save_csv = TRUE)
+drug_atc_codes(save_table = TRUE, save_csv = TRUE)
 
 # save parsed dataframe as csv if it does not exist in given location and
 # return parsed dataframe.
