+++ man/targets_links.Rd
@@ -52,7 +52,7 @@ targets_links(save_csv = TRUE)
 # save in database, save parsed dataframe as csv if it does not
 # exist in current location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-targets_links(ssave_table = TRUE, save_csv = TRUE)
+targets_links(save_table = TRUE, save_csv = TRUE)
 
 # save parsed dataframe as csv if it does not exist in given
 # location and return parsed dataframe.
