+++ man/transporters_articles.Rd
@@ -53,7 +53,7 @@ transporters_articles(save_csv = TRUE)
 # save in database, save parsed dataframe as csv if it does not exist
 # in current location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-transporters_articles(ssave_table = TRUE, save_csv = TRUE)
+transporters_articles(save_table = TRUE, save_csv = TRUE)
 
 # save parsed dataframe as csv if it does not exist in given location
 # and return parsed dataframe.
