+++ man/enzymes_polypeptide_go.Rd
@@ -53,7 +53,7 @@ enzymes_polypeptide_go(save_csv = TRUE)
 #  current location and return parsed dataframe.
 # If the csv exist before read it and return its data.
 enzymes_polypeptide_go(
-  ssave_table = TRUE,
+  save_table = TRUE,
   save_csv = TRUE
 )
 
