+++ man/drug_external_links.Rd
@@ -51,7 +51,7 @@ drug_external_links(save_csv = TRUE)
 # save in database, save parsed dataframe as csv if it does not exist in
 # current location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-drug_external_links(ssave_table = TRUE, save_csv = TRUE)
+drug_external_links(save_table = TRUE, save_csv = TRUE)
 
 # save parsed dataframe as csv if it does not exist in given location
 # and return parsed dataframe.
