+++ man/targets.Rd
@@ -50,7 +50,7 @@ targets(save_csv = TRUE)
 # save in database, save parsed dataframe as csv if it does not
 # exist in current location and return parsed dataframe.
 # If the csv exist before read it and return its data.
-targets(ssave_table = TRUE, save_csv = TRUE)
+targets(save_table = TRUE, save_csv = TRUE)
 
 # save parsed dataframe as csv if it does not exist in given
 # location and return parsed dataframe.
