+++ vignettes/Database Saving.Rmd
@@ -4,7 +4,7 @@ author: "Mohammed Ali"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{DrugBank Parser}
+  %\VignetteIndexEntry{Saving Data into Database}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
