### 2020-01-03 2e7246c
#72, fix build error
