### 2020-01-03 3d53243
#72, remove space from doc file
