### 2020-01-06 bf981b0
#14, improve package help page
