+++ man/dbparser.Rd
@@ -7,6 +7,14 @@
 \title{dbparser: A package for reading and parsing \strong{DrugBank} xml with the
 option to save it in a given db.}
 \description{
+The main purpose of the `dbparser` package is to parse the
+\strong{DrugBank} (\link{http://drugbank.ca}) database which is downloadable
+in XML format from \link{https://www.drugbank.ca/releases/latest}.
+The parsed data can then be explored and analyzed as desired by the user.
+In addition to that user can save parsed dataframe into desired database as
+well.
+}
+\details{
 dbparser package provides three categories of important functions:
 database related methods, xml db reader and \strong{DrugBank} elements
  parsers.
