### 2020-01-17 b1b7898
some edits to the package readme
