+++ README.md
@@ -28,33 +28,33 @@ accessible, online database containing information on drugs and drug
 targets. As both a bioinformatics and a cheminformatics resource,
 DrugBank combines detailed drug (i.e. chemical, pharmacological and
 pharmaceutical) data with comprehensive drug target (i.e. sequence,
-structure, and pathway) information, more information about DrugBank can
+structure, and pathway) information. More information about DrugBank can
 be found [here](https://www.drugbank.ca/about/).
 
-Using DrugBank database is not an easy task. As it is stored in a single
-XML file that can be viewed using XMLviewer as stated
-[here](http://shorturl.at/erEM7).
+In its raw form, the DrugBank database is a single XML file. Users must
+create an [account](https://www.drugbank.ca/public_users/sign_up) with
+Drugbank and request permission to
+[download](https://www.drugbank.ca/releases/latest) the database. Note
+that this may take a couple of days.
 
-Here come the importance of `dbparser` package that parses DrugBank
-database into different `R` tibbles that can be then be explored and
-analyzed as desired by the user, check this
-[tutorial](https://dainanahan.github.io/dbparser/articles/dbparser.html).
+The `dbparser` package parses the DrugBank XML database into `R` tibbles
+to then be explored and analyzed by the user. The package also offers
+the option to save these tibbles in databases including **SQL Server
+DDB** and **Maria DB**. See our
+[tutorial](https://dainanahan.github.io/dbparser/articles/dbparser.html)
+to get started.
 
-Moreover; user can save these tibbles in different databases like **SQL
-Server DDB** and **Maria DB** using `dbparser`.
+If you are waiting for access to the DrugBank database, or do not intend
+to do a deep dive with the data, you may wish to use the `dbdataset`
+[package](https://dainanahan.github.io/dbdataset/index.html), which
+contains the DrugBank database already parsed into `R` tibbles. Note
+that this is a large package that exceeds the limit set by CRAN. It is
+only available on GitHub.
 
 `dbparser` is tested against DrugBank versions *5.1.0* through *5.1.4*
-successfully. If you found errors with these versions or any other
+successfully. If you find errors with these versions or any other
 version please submit an issue
 [here](https://github.com/Dainanahan/dbparser/issues).
-
-Final note, in order to download DrugBank database you have to create a
-count [here](https://www.drugbank.ca/public_users/sign_up) and wait a
-little for registration confirmation.
-
-While you are waiting you can make use from `dbdataset` data package
-[here](https://dainanahan.github.io/dbdataset/index.html). It contains
-parsed DrugBank database `R` tibbles using `dbparser`.
 
 ## Installation
 
@@ -95,8 +95,8 @@ citation("dbparser")
 #> 
 #> To cite dbparser in publications use:
 #> 
-#>   Mohammed Ali, Ali Ezzat (). dbparser: DrugBank Database XML
-#>   Parser. R package version 1.0.5.9000.
+#>   Mohammed Ali, Ali Ezzat (). dbparser: DrugBank Database XML Parser. R
+#>   package version 1.0.5.9000.
 #> 
 #> A BibTeX entry for LaTeX users is
 #> 
