### 2020-01-18 #347 @emmamendelsohn
Hi @MohammedFCIS! Nice job with the updates--the package feels much more polished. I think the new function names work well.

I like that you shortened the README and made a separate tutorial. I'm submitting a PR with a few suggested changes to the README text. Feel free to use or discard as you wish :)

I have just a few comments on this second review. Let me know if you need any clarification.

**Database capabilities**
- I see that you added a vignette on this. Can you link to it from the README and tutorial so that the user can easily locate it?
- Echoing my comments from the first review: is it necessary for the user to connect to a server? What if the user wants to save to a local `DBI` connection?

**Tutorial**
- I would add something to the top saying that it assumes you have already downloaded the data (per instructions in the README), saved it to your working directory, and renamed it to `drugbank.xml`.
- Show options to save data in your functions (e.g., `drug(save_csv=TRUE)`).
- The function hyperlinks are broken.

**Dataframes versus tibbles**
- Sorry I missed this the first time: some functions return dataframes (e.g., `drug_articles()`) but most return tibbles. I think from your package description you intend for everything to return tibbles, so you may need to do a sweep to make sure all are tibbles. In addition, the reference page is inconsistent about reporting tibble versus dataframe, so it should be updated too.

**Parsers**
- When running a function without first running `read_drugbank_xml_db("drugbank.xml")`, an empty tibble is returned. Maybe add a message reminding users to first run the read function?

**R help**
- I know it's not a major point of entry to the package, but I'd suggest updating the `?dbparser` text to match some of the language you developed in the README etc.
