### 2020-01-18 #347 @MohammedFCIS
Hello @emmamendelsohn ,
As usual I am enjoying your notes and your contribution is very welcome and appreciated of course :).
I have created issues with your notes here https://github.com/Dainanahan/dbparser/labels/ropensci and I will work on it immediately.
I just need clarifications about these two points:
* Show options to save data in your functions (e.g., drug(save_csv=TRUE)) ==> Show it where as it is written in the documentation?
* The function hyperlinks are broken ==> (what functions links are broken).
For *DBI* part I will clarify it more on the tutorial, but user can use anything that DBI support including local db.
Thank you.
