### 2020-01-19 9fd899f
#80, make sure all returned data is of type tibble
