+++ tests/testthat/test_drug_main_node_parser_biotech.R
@@ -23,6 +23,7 @@ test_that(
   desc = "Read darug primary key attribute",
   code = {
     expect_match(drug()[1][["primary_key"]], "DB00001")
+    expect_true(is_tibble(drug()))
     expect_error(drug(TRUE))
   }
 )
@@ -42,6 +43,7 @@ test_that(
       as.character(drug_groups()[1][["group"]]),
       "approved"
     )
+    expect_true(is_tibble(drug_groups()))
     expect_error(drug_groups(TRUE))
   }
 )
@@ -53,6 +55,7 @@ test_that(
       as.character(drug_articles()[["pubmed-id"]][1]),
       "16244762"
     )
+    expect_true(is_tibble(drug_articles()))
     expect_error(drug_articles(TRUE))
   }
 )
@@ -61,6 +64,7 @@ test_that(
   desc = "Read darug books attributes",
   code = {
     expect_equal(nrow(drug_books()), 0)
+    expect_true(is_tibble(drug_books()))
     expect_error(drug_books(TRUE))
   }
 )
@@ -72,6 +76,7 @@ test_that(
       as.character(drug_links()[1][["title"]]),
       "Google books"
     )
+    expect_true(is_tibble(drug_links()))
     expect_error(drug_links(TRUE))
   }
 )
@@ -83,6 +88,7 @@ test_that(
       drug_classification()[1][["parent_key"]],
       "DB00001"
     )
+    expect_true(is_tibble(drug_classification()))
     expect_error(drug_classification(TRUE))
   }
 )
@@ -94,6 +100,7 @@ test_that(
       drug_syn()[["synonym"]][[1]],
       "Hirudin variant-1"
     )
+    expect_true(is_tibble(drug_syn()))
     expect_error(drug_syn(TRUE))
   }
 )
@@ -105,6 +112,7 @@ test_that(
       as.character(drug_articles()[["pubmed-id"]][1]),
       "16244762"
     )
+    expect_true(is_tibble(drug_articles()))
     expect_error(drug_articles(TRUE))
   }
 )
@@ -116,6 +124,7 @@ test_that(
       as.character(drug_products()[["name"]][1]),
       "Refludan"
     )
+    expect_true(is_tibble(drug_products()))
     expect_error(drug_products(TRUE))
   }
 )
@@ -127,6 +136,7 @@ test_that(
       as.character(drug_mixtures()[["name"]][1]),
       "Refludan"
     )
+    expect_true(is_tibble(drug_mixtures()))
     expect_error(drug_mixtures(TRUE))
   }
 )
@@ -138,6 +148,7 @@ test_that(
       as.character(drug_packagers()[["name"]][1]),
       "Bayer Healthcare"
     )
+    expect_true(is_tibble(drug_packagers()))
     expect_error(drug_packagers(TRUE))
   }
 )
@@ -149,6 +160,7 @@ test_that(
       as.character(drug_manufacturers()[["manufacturer"]][[1]]),
       "Bayer healthcare pharmaceuticals inc"
     )
+    expect_true(is_tibble(drug_manufacturers()))
     expect_error(drug_manufacturers(TRUE))
   }
 )
@@ -160,6 +172,7 @@ test_that(
       as.character(drug_prices()[["currency"]][[1]]),
       "USD"
     )
+    expect_true(is_tibble(drug_prices()))
     expect_error(drug_prices(TRUE))
   }
 )
@@ -171,6 +184,7 @@ test_that(
       as.character(drug_categories()[["mesh-id"]][[1]]),
       "D000602"
     )
+    expect_true(is_tibble(drug_categories()))
     expect_error(drug_categories(TRUE))
   }
 )
[…]
