+++ tests/testthat/test_drug_enzyme_node_parser_biotech.R
@@ -25,6 +25,7 @@ test_that(
       nrow(enzymes()),
       0
     )
+    expect_true(is_tibble(enzymes()))
     expect_error(enzymes(TRUE))
   }
 )
@@ -36,6 +37,7 @@ test_that(
       nrow(enzymes_actions()),
       0
     )
+    expect_true(is_tibble(enzymes_actions()))
     expect_error(enzymes_actions(TRUE))
   }
 )
@@ -47,6 +49,7 @@ test_that(
       nrow(enzymes_articles()),
       0
     )
+    expect_true(is_tibble(enzymes_articles()))
     expect_error(enzymes_articles(TRUE))
   }
 )
@@ -59,6 +62,7 @@ test_that(
       nrow(enzymes_textbooks()),
       0
     )
+    expect_true(is_tibble(enzymes_textbooks()))
     expect_error(enzymes_textbooks(TRUE))
   }
 )
@@ -70,6 +74,7 @@ test_that(
       nrow(enzymes_links()),
       0
     )
+    expect_true(is_tibble(enzymes_links()))
     expect_error(enzymes_links(TRUE))
   }
 )
@@ -81,6 +86,7 @@ test_that(
       nrow(enzymes_polypeptide()),
       0
     )
+    expect_true(is_tibble(enzymes_polypeptide()))
     expect_error(enzymes_polypeptide(TRUE))
   }
 )
@@ -93,6 +99,7 @@ test_that(
       nrow(enzymes_polypeptide_ext_ident()),
       0
     )
+    expect_true(is_tibble(enzymes_polypeptide_ext_ident()))
     expect_error(
       enzymes_polypeptide_ext_ident(TRUE)
     )
@@ -107,6 +114,7 @@ test_that(
       nrow(enzymes_polypeptide_syn()),
       0
     )
+    expect_true(is_tibble(enzymes_polypeptide_syn()))
     expect_error(enzymes_polypeptide_syn(TRUE))
   }
 )
@@ -118,6 +126,7 @@ test_that(
       nrow(enzymes_polypeptide_pfams()),
       0
     )
+    expect_true(is_tibble(enzymes_polypeptide_pfams()))
     expect_error(enzymes_polypeptide_pfams(TRUE))
   }
 )
@@ -129,6 +138,7 @@ test_that(
       nrow(enzymes_polypeptide_go()),
       0
     )
+    expect_true(is_tibble(enzymes_polypeptide_go()))
     expect_error(enzymes_polypeptide_go(TRUE))
   }
 )
