+++ tests/testthat/test_drug_reaction_node_parser_biotech.R
@@ -23,6 +23,7 @@ test_that(
       nrow(drug_reactions()),
       0
     )
+    expect_true(is_tibble(drug_reactions()))
     expect_error(drug_reactions(TRUE))
   }
 )
@@ -34,6 +35,7 @@ test_that(
       nrow(drug_reactions_enzymes()),
       0
     )
+    expect_true(is_tibble(drug_reactions_enzymes()))
     expect_error(drug_reactions_enzymes(TRUE))
   }
 )

[…]
