### 2020-01-19 4e0996a
#80 update function documentation
