+++ R/dbparser.R
@@ -5,7 +5,7 @@
 #' \strong{DrugBank} database which is downloadable
 #' in XML format.
 #' The parsed data can then be explored and analyzed as desired by the user.
-#' In addition to that user can save parsed dataframe into desired database as
+#' In addition to that user can save parsed tibble into desired database as
 #' well.
 #'
 #' dbparser package provides three categories of important functions:
@@ -22,7 +22,7 @@
 #'
 #' @section parsers functions:
 #'  Each parser function is responsible of parsing certain drug element and
-#'  returning its dataframe
+#'  returning its tibble
 #'  with the ability to save it in a predefined database.
 #'
 #' @docType package

[…]
