### 2020-01-19 0b10ddc
fix #80, update related documentation
