### 2020-01-19 e284ec5
fix #81, check if 'read_drugbank_xml_db' method is called before parsers
