+++ R/drug_atc_codes_node_parser.R
@@ -78,7 +78,7 @@ get_atc_codes_df <- function(rec) {
 #' @export
 drug_atc_codes <- function(save_table = FALSE, save_csv = FALSE,
                                  csv_path = ".", override_csv = FALSE) {
-  check_database_connection(save_table)
+  check_data_and_connection(save_table)
   path <- get_dataset_full_path("drug_atc_codes", csv_path)
   if (!override_csv & file.exists(path)) {
     drug_atc_codes <- readr::read_csv(path)
