+++ R/drug_parsers.R
@@ -53,7 +53,7 @@ drug <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
+    check_data_and_connection(save_table)
     path <- get_dataset_full_path("drugs", csv_path)
     if (!override_csv & file.exists(path)) {
       drugs <- readr::read_csv(path)
@@ -204,7 +204,7 @@ drug_groups <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
+    check_data_and_connection(save_table)
     path <- get_dataset_full_path("drug_groups", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_groups <- readr::read_csv(path)
@@ -284,7 +284,7 @@ drug_articles <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
+    check_data_and_connection(save_table)
     path <- get_dataset_full_path("drug_articles", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_articles <- readr::read_csv(path)
@@ -365,7 +365,7 @@ drug_books <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
+    check_data_and_connection(save_table)
     path <- get_dataset_full_path("drug_books", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_books <- readr::read_csv(path)
@@ -444,7 +444,7 @@ drug_links <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
+    check_data_and_connection(save_table)
     path <- get_dataset_full_path("drug_links", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_links <- readr::read_csv(path)
@@ -524,7 +524,7 @@ drug_syn <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
+    check_data_and_connection(save_table)
     path <- get_dataset_full_path("drug_syn", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_syn <- readr::read_csv(path)
@@ -601,7 +601,7 @@ drug_products <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
+    check_data_and_connection(save_table)
     path <- get_dataset_full_path("drug_products", csv_path)
     if (!override_csv & file.exists(path)) {
       drug_products <- readr::read_csv(path)
@@ -681,7 +681,7 @@ drug_calc_prop <- function(save_table = FALSE,
                                  save_csv = FALSE,
                                  csv_path = ".",
                                  override_csv = FALSE) {
-  check_database_connection(save_table)
+  check_data_and_connection(save_table)
   path <-
     get_dataset_full_path("drug_calculated_properties", csv_path)
   if (!override_csv & file.exists(path)) {
@@ -767,7 +767,7 @@ drug_intern_brand <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
+    check_data_and_connection(save_table)
     path <-
       get_dataset_full_path("drug_international_brands", csv_path)
     if (!override_csv & file.exists(path)) {
@@ -848,7 +848,7 @@ drug_salts <-
            save_csv = FALSE,
            csv_path = ".",
            override_csv = FALSE) {
-    check_database_connection(save_table)
[…]
