+++ R/drug_manufacturer_node_parser.R
@@ -1,20 +1,3 @@
-# Extract drug manufacturers df
-get_manufacturer_rec <- function(r, drug_key) {
-  tibble(
-    name = xmlValue(r),
-    url = xmlGetAttr(r, name = "url"),
-    generic = xmlGetAttr(r, name = "generic"),
-    parent_key = drug_key
-  )
-}
-
-get_manufactures_df <- function(rec) {
-  return(map_df(
-    xmlChildren(rec[["manufacturers"]]),
-    ~ get_manufacturer_rec(., xmlValue(rec["drugbank-id"][[1]]))
-  ))
-}
-
 #' Extracts the drug manufacturers element and return data as tibble.
 #'
 #' \code{drug_manufacturers} returns tibble of drug manufacturers
@@ -69,7 +52,7 @@ get_manufactures_df <- function(rec) {
 #' @export
 drug_manufacturers <- function(save_table = FALSE, save_csv = FALSE,
                                      csv_path = ".", override_csv = FALSE) {
-  check_database_connection(save_table)
+  check_data_and_connection(save_table)
   path <-
     get_dataset_full_path("drug_manufacturers", csv_path)
   if (!override_csv & file.exists(path)) {
