### 2020-01-19 9ca4155
#78, add sql lite support and enahce unit tests
