+++ NAMESPACE
@@ -104,6 +104,7 @@ importFrom(purrr,map_chr)
 importFrom(purrr,map_df)
 importFrom(readr,read_csv)
 importFrom(readr,write_csv)
+importFrom(tibble,as_tibble)
 importFrom(tibble,tibble)
 importFrom(tidyr,'%>%')
 importFrom(utils,unzip)
