### 2020-01-19 e7725ee
#78, add RSQLite dependency
