### 2020-01-20 #347 @emmamendelsohn
- Yes, a very minor suggestion: you may want to mention in the tutorial that the user has the option to set `save_csv=TRUE` in the parse functions.
- Hmm, this may actually be beyond your control? When I view the tutorial, all the functions are hyperlinks. Clicking on `drug()` takes me to this link, which is not found: https://rdrr.io/cran/dbparser/man/drug.html. I'm guessing it has to do with the GitHub version falling out of sync with CRAN?
