### 2020-01-20 #347 @MohammedFCIS
Thank you @emmamendelsohn for clarifications.
For the links point, I think you are right it is out of my hand. What I am using are the references from package site here https://dainanahan.github.io/dbparser/reference/index.html and https://dainanahan.github.io/dbparser/reference/drug.html
