### 2020-01-20 37acc81
fix #78, enahance database tutorial.
