+++ README.Rmd
@@ -44,7 +44,7 @@ the database. Note that this may take a couple of days.
 
 The `dbparser` package parses the DrugBank XML database into `R` tibbles that can be explored and analyzed by the user, check [this tutorial](https://dainanahan.github.io/dbparser/articles/dbparser.html) for more details.
 
-Also, the package offers the option to save these tibbles in databases including **SQL Server DDB** and **Maria DB** just by enabling `save_table` option,  check [this tutorial](https://dainanahan.github.io/dbparser/articles/Database_Saving.html) to get started.
+Also, the package offers the option to save these tibbles in databases including **SQL Server DDB** and **Maria DB** just by enabling `save_table` option,  check [this tutorial](https://dainanahan.github.io/dbparser/articles/Database_Saving.html) for more details.
 
 If you are waiting for access to the DrugBank database, or do not intend to do a deep dive with
 the data, you may wish to use the `dbdataset`
