+++ vignettes/dbparser.Rmd
@@ -28,8 +28,15 @@ In this tutorial, we will see how to use `dbparser` along with `dplyr` and
 
 ## Loading and Parsing the Data
 
-Following is sample code that parses the DrugBank database, then loads the 
-drugs info, drug groups info and drug targets actions info.
+Before starting the code we are assuming the following:
+
+- user already downloaded *DrugBank* xml database file based on the
+[Read Me](https://dainanahan.github.io/dbparser) instructions or the above note,
+- user saved the downloaded database in working directory as `C:\`.
+- user named the downloaded xml file **drugbank.xml**. 
+
+Now we can loads the `drugs` info, `drug groups` info and `drug targets`
+actions info.
 
 ```{r eval=FALSE}
 ## load dbparser package
@@ -39,7 +46,7 @@ library(ggplot2)
 library(XML)
 
 ## parse data from XML and save it to memory
-read_drugbank_xml_db("drugbank.xml")
+read_drugbank_xml_db("C:\drugbank.xml")
 
 ## load drugs data
 drugs <- drug()
