### 2020-01-22 #347 @MohammedFCIS
Hello @emmamendelsohn ,
I finished your notes, kindly check here https://github.com/Dainanahan/dbparser/labels/ropensci.
Thank you.
