### 2020-01-31 #347 @emmamendelsohn
Thanks for the reminder, @MohammedFCIS. I reviewed your responses to my comments and everything looks good. I just submitted a PR with some edits to the database vignette. @noamross it would be helpful if you could give the vignette a read-through as well.
