### 2020-01-31 04d6b0e
Update drug_common_utilities.R
