### 2020-02-01 #347 @haozhu233
Hi, all, sorry for being a little late on getting back on this. I had been quite busy recently.

@MohammedFCIS I was actually amazed by the amount of updates you made since I reviewed it last time. Right now this package looks very polished and I would say the software interface looks very clean right now and very user friendly! Impressive!

After reviewing this new version several times, I only have a tiny comment regarding the use of `%>%`. I noticed that you imported pipe from `tidyr` and that is the only thing you imported from that package. This is a little weird. You can either imported it from `magrittr` or from `purrr`, which also has `%>%` exported.
