### 2020-02-02 Dainanahan/dbparser#85 @MohammedFCIS
Resolve migratter issue
After reviewing this new version several times, I only have a tiny comment regarding the use of %>%. I noticed that you imported pipe from tidyr and that is the only thing you imported from that package. This is a little weird. You can either imported it from magrittr or from purrr, which also has %>% exported.
