+++ R/imports.R
@@ -1,10 +1,9 @@
 #' @importFrom DBI dbWriteTable dbExecute dbConnect dbDisconnect
 #' @importFrom odbc odbc
-#' @importFrom purrr map_df map_chr
+#' @importFrom purrr '%>%' map_df map_chr
 #' @importFrom readr read_csv write_csv
 #' @importFrom RMariaDB MariaDB
 #' @importFrom tibble tibble as_tibble
-#' @importFrom tidyr '%>%'
 #' @importFrom utils unzip
 #' @importFrom XML xmlChildren xmlGetAttr xmlValue xmlToList xmlToDataFrame
 #' xmlParse xmlRoot xmlSize xmlApply
