+++ NAMESPACE
@@ -100,11 +100,11 @@ importFrom(XML,xmlToDataFrame)
 importFrom(XML,xmlToList)
 importFrom(XML,xmlValue)
 importFrom(odbc,odbc)
+importFrom(purrr,'%>%')
 importFrom(purrr,map_chr)
 importFrom(purrr,map_df)
 importFrom(readr,read_csv)
 importFrom(readr,write_csv)
 importFrom(tibble,as_tibble)
 importFrom(tibble,tibble)
-importFrom(tidyr,'%>%')
 importFrom(utils,unzip)
