### 2020-02-02 #347 @MohammedFCIS
Thank you so much @haozhu233 for your feedback :).  It is only due to your and @emmamendelsohn  very useful notes ;).
I have fixed your note here https://github.com/Dainanahan/dbparser/issues/85
Best Regards,
