### 2020-02-04 #347 @noamross
Thank you @emmamendelsohn and @haozhu233 for your responses.  I am doing final editor checks and have a an issue on the database vignette / `open_db()` function and documentation.  The function is hard-coded for either SQLite or an _ODBC_ database, but this isn't clear in the vignette, which only refers to "DBI-compliant" databases.  While this could be fixed by documentation, I think it would be simpler for instead the `open_db()` function to simply take a _connection_ as its first argument, so you don't have to hard-code any of these options.  That is, `open_db(RSQLite::RSQLite(...))` passes this connection to `pkg_env$con`.  The default or `open_mdb` can still go to MariaDB.

-
