### 2020-02-12 d97927b
#87 change methods parameters and related docs
