+++ R/drug_reaction_node_parser.R
@@ -30,8 +30,9 @@ get_reactions_enzymes_df <- function(rec) {
 #'
 #' This functions extracts the groups element of drug node in \strong{DrugBank}
 #' xml database with the option to save it in a predefined database via
-#' \code{\link{open_db}} method. It takes one single optional argument to
-#' save the returned tibble in the database.
+#' passed database connection. It takes two optional arguments to
+#' save the returned tibble in the database \code{save_table} and
+#' \code{database_connection}.
 #' It must be called after \code{\link{read_drugbank_xml_db}} function like
 #' any other parser function.
 #' If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
@@ -43,6 +44,10 @@ get_reactions_enzymes_df <- function(rec) {
 #' location, save_csv must be true
 #' @param override_csv override existing csv, if any, in case it is true in the
 #'  new parse operation
+#' @param database_connection DBI connection object that holds a connection to
+#' user defined database. If \code{save_table} is enabled without providing
+#' value for this function an error will be thrown.
+#'
 #' @return drug reactions node attributes date frame
 #'
 #' @examples
@@ -50,8 +55,12 @@ get_reactions_enzymes_df <- function(rec) {
 #' # return only the parsed tibble
 #' drug_reactions()
 #'
-#' # save in database and return parsed tibble
+#' # will throw an error, as database_connection is NULL
 #' drug_reactions(save_table = TRUE)
+#'
+#' # save in database in SQLite in memory database and return parsed tibble
+#' sqlite_con <- DBI::dbConnect(RSQLite::SQLite())
+#' drug_reactions(save_table = TRUE, database_connection = sqlite_con)
 #'
 #' # save parsed tibble as csv if it does not exist in current location
 #' # and return parsed tibble.
@@ -61,7 +70,8 @@ get_reactions_enzymes_df <- function(rec) {
 #' # save in database, save parsed tibble as csv if it does not exist
 #' # in current location and return parsed tibble.
 #' # If the csv exist before read it and return its data.
-#' drug_reactions(save_table = TRUE, save_csv = TRUE)
+#' drug_reactions(save_table = TRUE, save_csv = TRUE,
+#'  database_connection = sqlite_con)
 #'
 #' # save parsed tibble as csv if it does not exist in given location
 #' # and return parsed tibble.
@@ -78,8 +88,9 @@ drug_reactions <-
   function(save_table = FALSE,
            save_csv = FALSE,
            csv_path = ".",
-           override_csv = FALSE) {
-    check_data_and_connection(save_table)
+           override_csv = FALSE,
+           database_connection = NULL) {
+    check_parameters_validation(save_table, database_connection)
     path <-
       get_dataset_full_path("drug_reactions", csv_path)
     if (!override_csv & file.exists(path)) {
@@ -94,7 +105,7 @@ drug_reactions <-
 
     if (save_table) {
       save_drug_sub(
-        con = pkg_env$con,
+        con = database_connection,
         df = drug_reactions,
         table_name = "drug_reactions",
         foreign_key = "parent_key"
@@ -111,8 +122,9 @@ drug_reactions <-
 #' This functions extracts the reactions enzymes element of drug node in
 #'  drugbank
 #' xml database with the option to save it in a predefined database via
-#' \code{\link{open_db}} method. It takes one single optional argument to
-#' save the returned tibble in the database.
+#' passed database connection. It takes two optional arguments to
+#' save the returned tibble in the database \code{save_table} and
+#' \code{database_connection}.
 #' It must be called after \code{\link{read_drugbank_xml_db}} function like
[…]
