### 2020-02-12 45fa475
#87 change methods parameters and related docs
