### 2020-02-13 ded79a4
#87, fix unit tests
