+++ tests/testthat/test_drug_carrier_node_parser_biotech.R
@@ -132,3 +132,4 @@ test_that(
     expect_error(carriers_polypeptides_pfams(TRUE))
   }
 )
+dbDisconnect(database_connection)
