### 2020-02-14 57aa3ec
#87, do not save empty tibble as it causes database errors
