+++ NAMESPACE
@@ -98,6 +98,7 @@ importFrom(XML,xmlToList)
 importFrom(XML,xmlValue)
 importFrom(odbc,odbc)
 importFrom(purrr,'%>%')
+importFrom(purrr,is_empty)
 importFrom(purrr,map_chr)
 importFrom(purrr,map_df)
 importFrom(readr,read_csv)
