### 2020-02-14 dc33b02
#87. add database unit tests
