### 2020-02-14 b94b97e
#87, update manufacturers documentation
