### 2020-02-14 32c3dec
#87, update reference documentation
