### 2020-02-14 ca54483
fix #87. update database tutorial
