### 2020-02-14 #347 @MohammedFCIS
Thank you @noamross so much for your valuable note. It inspired me to change a lot actually. Please find the fix in that issue https://github.com/Dainanahan/dbparser/issues/87.
I have updated both the code, documentation and tutorial.
@haozhu233, @emmamendelsohn please feel free to review this update as well.
Thank you all for your time and valuable input.
