### 2020-02-18 #347 @noamross
Thank you, @MohammedFCIS, for the update.  The package is approved! Thanks for submitting and @emmamendelsohn and @haozhu233 for your reviews!

To-dos:
- [ ] If you already had a `pkgdown` website, fix its URL to point to `https://docs.ropensci.org/package_name` and deactivate the automatic deployment you might have set up, since it will not be built centrally like for all rOpenSci packages, see http://devdevguide.netlify.com/#docsropensci. In addition, in your DESCRIPTION file, include the docs link in the `URL` field alongside the link to the GitHub repository, e.g.: `URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar`
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL.

Should you want to acknowledge your reviewers in your package DESCRIPTION, you can do so by making them `"rev"`-type contributors in the `Authors@R` field.  More info on this [here](https://devguide.ropensci.org/building.html#authorship).

We've put together an online book with our best practice and tips, [this chapter](https://devguide.ropensci.org/collaboration.html) starts the  section that's about guidance for post-review. Please tell us what could be improved, the corresponding repo is [here](https://github.com/ropensci/dev_guide).
