+++ man/dbparser.Rd
@@ -24,7 +24,7 @@ To achieve this purpose, `dbparser`` package provides three main categories
 - and database related methods.
 
 For more infomration kindly check the
-reference/index (https://dainanahan.github.io/dbparser/reference/index.html)
+reference/index (https://ropensci.github.io/dbparser/reference/index.html)
 }
 \section{xml db reader functions}{
 
