+++ CONTRIBUTING.md
@@ -6,11 +6,11 @@ First of all, thanks for considering contributing to drugverse! 👍 It's people
 
 drugverse is an open source project, maintained by people who care. We are not directly funded to do so.
 
-[repo]: https://github.com/Dainanahan/dbparser
-[issues]: https://github.com/Dainanahan/dbparser/issues
-[new_issue]: https://github.com/Dainanahan/dbparser/issues/new
-[website]: https://dainanahan.github.io/dbparser
-[citation]: https://dainanahan.github.io/dbparser/authors.html
+[repo]: https://github.com/ropensci/dbparser
+[issues]: https://github.com/ropensci/dbparser/issues
+[new_issue]: https://github.com/ropensci/dbparser/issues/new
+[website]: https://ropensci.github.io/dbparser
+[citation]: https://ropensci.github.io/dbparser/authors.html
 [email]: mohammed.ali@edu.dsti.institute
 
 ## Code of conduct
