+++ vignettes/dbparser.Rmd
@@ -31,7 +31,7 @@ In this tutorial, we will see how to use `dbparser` along with `dplyr` and
 Before starting the code we are assuming the following:
 
 - user already downloaded *DrugBank* xml database file based on the
-[Read Me](https://dainanahan.github.io/dbparser) instructions or the above note,
+[Read Me](https://ropensci.github.io/dbparser) instructions or the above note,
 - user saved the downloaded database in working directory as `C:\`.
 - user named the downloaded xml file **drugbank.xml**. 
 
@@ -58,7 +58,7 @@ drug_groups <- drug_groups()
 drug_targets_actions <- targets_actions()
 ```
 ### Saving into Database
-User can save parsed tibbles in database by enabling `save_table` option. More details about that can be found in [database tutorial](https://dainanahan.github.io/dbparser/articles/Database_Saving.html)
+User can save parsed tibbles in database by enabling `save_table` option. More details about that can be found in [database tutorial](https://ropensci.github.io/dbparser/articles/Database_Saving.html)
 
 ## Exploring the data
 
