+++ README.Rmd
@@ -42,20 +42,20 @@ XML file. Users must create an [account](https://www.drugbank.ca/public_users/si
 with Drugbank and request permission to [download](https://www.drugbank.ca/releases/latest)
 the database. Note that this may take a couple of days.
 
-The `dbparser` package parses the DrugBank XML database into `R` tibbles that can be explored and analyzed by the user, check [this tutorial](https://dainanahan.github.io/dbparser/articles/dbparser.html) for more details.
+The `dbparser` package parses the DrugBank XML database into `R` tibbles that can be explored and analyzed by the user, check [this tutorial](https://ropensci.github.io/dbparser/articles/dbparser.html) for more details.
 
-Also, the package offers the option to save these tibbles in databases including **SQL Server DDB** and **Maria DB** just by enabling `save_table` option,  check [this tutorial](https://dainanahan.github.io/dbparser/articles/Database_Saving.html) for more details.
+Also, the package offers the option to save these tibbles in databases including **SQL Server DDB** and **Maria DB** just by enabling `save_table` option,  check [this tutorial](https://ropensci.github.io/dbparser/articles/Database_Saving.html) for more details.
 
 If you are waiting for access to the DrugBank database, or do not intend to do a deep dive with
 the data, you may wish to use the `dbdataset`
-[package](https://dainanahan.github.io/dbdataset/index.html), which contains
+[package](https://ropensci.github.io/dbdataset/index.html), which contains
 the DrugBank database already parsed into `R` tibbles. Note that this is a large package that
 exceeds the limit set by CRAN. It is only available on GitHub.
 
 `dbparser` is tested against DrugBank versions *5.1.0* through *5.1.4*
 successfully. If you find errors with these versions or any other
 version please submit an issue
-[here](https://github.com/Dainanahan/dbparser/issues).
+[here](https://github.com/ropensci/dbparser/issues).
 
 ## Installation
 
@@ -69,7 +69,7 @@ or you can install the latest updates directly from the repo
 
 ``` r
 library(devtools)
-devtools::install_github("Dainanahan/dbparser")
+devtools::install_github("ropensci/dbparser")
 ```
 
 ## Code of Conduct
