+++ DESCRIPTION
@@ -30,5 +30,5 @@ Suggests:
     rmarkdown,
     testthat
 VignetteBuilder: knitr
-URL: https://dainanahan.github.io/dbparser/index.html
-BugReports: https://github.com/Dainanahan/dbparser/issues
+URL: https://github.com/ropensci/dbparser.git
+BugReports: https://github.com/ropensci/dbparser/issues
