### 2020-02-19 85a6f00
#88, Fix any links in badges for CI and coverage to point to the ropensci URL
