### 2020-02-19 ed9699a
#88, deactivate the automatic deployment for pkgdown site
