### 2020-02-19 e8c878c
#88, nclude the docs link in the URL field alongside the link to the GitHub repository
