### 2020-02-19 c91b7f8
fix #88, stop site autodeployment
