+++ .travis.yml
@@ -14,12 +14,3 @@ after_success:
   - Rscript -e 'library(pkgdown); build_site()'
 
 before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
-
-deploy:
-  provider: pages
-  skip-cleanup: true
-  github-token: $GITHUB_PAT
-  keep-history: true
-  local-dir: docs
-  on:
-    branch: master
