### 2020-02-19 #347 @MohammedFCIS
Hi @noamross, the following ToDo list is finished.
 If you already had a pkgdown website, fix its URL to point to https://docs.ropensci.org/package_name and deactivate the automatic deployment you might have set up, since it will not be built centrally like for all rOpenSci packages, see http://devdevguide.netlify.com/#docsropensci. In addition, in your DESCRIPTION file, include the docs link in the URL field alongside the link to the GitHub repository, e.g.: URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar
- [x] Fix any links in badges for CI and coverage to point to the ropensci URL.
- [x] Adding reviewers to package descriptions.

Thank you
